### 2023-08-01 #5725 @editorialbot
[PRE REVIEW]: StormR: An R package to quantify and map the tropical storms and cyclones’ winds characteristics
**Submitting author:** @thomasarsouze (<a href="http://orcid.org/0000-0002-8871-6120">Thomas Arsouze</a>)
**Repository:** https://github.com/umr-amap/StormR
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
