+++ paper/paper.bib
@@ -143,6 +143,7 @@ Publisher: Nature Publishing Group},
 	copyright = {Copyright Taylor and Francis Group, LLC},
 	issn = {0043-1672},
 	url = {https://www.tandfonline.com/doi/abs/10.1080/00431672.1974.9931702},
+	doi = {10.1080/00431672.1974.9931702},
 	abstract = {(1974). The Hurricane Disaster—Potential Scale. Weatherwise: Vol. 27, No. 4, pp. 169-186.},
 	language = {en},
 	urldate = {2023-06-05},
@@ -156,7 +157,7 @@ Publisher: Nature Publishing Group},
 @misc{knapp_international_2018,
 	title = {International {Best} {Track} {Archive} for {Climate} {Stewardship} ({IBTrACS}) {Project}, {Version} 4},
 	url = {https://www.ncei.noaa.gov/access/metadata/landing-page/bin/iso?id=gov.noaa.ncdc:C01552},
-	doi = {https://doi.org/10.25921/82ty-9e16},
+	doi = {10.25921/82ty-9e16},
 	abstract = {International Best Track Archive for Climate Stewardship (IBTrACS) Project, Version 4 (format: HTML)},
 	language = {en},
 	urldate = {2023-06-30},
