### 2023-08-08 #5725 @thomaspibanez
Hi @kthyng  @hugoledoux

I checked out and fixed the DOI notes.

Best wishes,

Thomas
