+++ pkgdown.yml
@@ -8,5 +8,5 @@ articles:
   Models: Models.html
   SpatialBehaviour: SpatialBehaviour.html
   TemporalBehaviour: TemporalBehaviour.html
-last_built: 2023-07-24T13:00Z
+last_built: 2023-08-08T08:24Z
 
