### 2023-08-08 #5725 @hugoledoux
@thomasarsouze while I am looking for reviewers, I noticed that 4 people contributed to the code on github, but there are 6 authors... JOSS deals with this differently from "traditional" journals, we don't consider the person funding the project as possible authors. See our rules there: https://joss.readthedocs.io/en/latest/submitting.html?highlight=authorship#authorship

Of course maybe those people contributed, but if you could write here what they did that would help us move further.
