### 2023-08-09 #5725 @thomaspibanez
@hugoledoux Swen Jullien and Christophe Menkes (listed here : https://umr-amap.github.io/StormR/) are two climatologist working on tropical cyclones. They participated to regular meeting regarding the development of the package, helped with cyclone models selection, and with the writing of the documentation. They didn't work on the project with GitHub that why they are not in the contributor list but their contribution was essential.
