### 2023-08-21 umr-amap/StormR#68 @highamm
[Bug]: Boose method
### What happened?

I am trying to use the `"Boose"` method in the `spatialBehaviour()` function. While the other methods (`"Willoughby"` and `"Holland"`) work well, I get an Error using `"Boose"`.

### Minimal Complete Verifiable Example

```R
library(StormR)
sds <- defStormsDataset()
st <- defStormsList(sds = sds, loi = c(168.33, -17.73),
                    names = "PAM", verbose = 0)

## Willoughby works
pf <- spatialBehaviour(st, product = "Profiles",
                       method = "Willoughby",
                       asymmetry = "Chen", verbose = 0)

## Boose does not
pf <- spatialBehaviour(st, product = "Profiles",
                       method = "Boose", verbose = 0)

## Error in wk_handle.wk_wkb(wkb, s2_geography_writer(oriented = oriented,  :
  Loop 83 is not valid: Edge 750 crosses edge 1005
```

### Propose a solution

_No response_

### Anything else we need to know?

Below is my session information.

```
> sessionInfo()
R version 4.2.1 (2022-06-23)
Platform: x86_64-apple-darwin17.0 (64-bit)
Running under: macOS Ventura 13.4

Matrix products: default
LAPACK: /Library/Frameworks/R.framework/Versions/4.2/Resources/lib/libRlapack.dylib

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

attached base packages:
[1] stats     graphics  grDevices utils     datasets  methods
[7] base

other attached packages:
[1] StormR_0.1.0

 [1] Rcpp_1.0.10        pillar_1.9.0       compiler_4.2.1
 [4] viridis_0.6.2      class_7.3-20       tools_4.2.1
 [7] dotCall64_1.0-2    ncdf4_1.21         viridisLite_0.4.1
[10] gtable_0.3.1       lifecycle_1.0.3    tibble_3.2.1
[13] lattice_0.20-45    pkgconfig_2.0.3    rlang_1.1.1
[16] DBI_1.1.3          cli_3.6.1          rstudioapi_0.14
[19] spam_2.9-1         gridExtra_2.3      e1071_1.7-11
[22] terra_1.7-29       dplyr_1.1.2        s2_1.1.0
[25] stringr_1.5.0      generics_0.1.3     vctrs_0.6.2
[28] fields_14.1        maps_3.4.1         classInt_0.4-8
[31] grid_4.2.1         tidyselect_1.2.0   glue_1.6.2
[34] sf_1.0-9           R6_2.5.1           fansi_1.0.4
[37] foreign_0.8-82     rworldmap_1.3-6    sp_1.5-1
[40] ggplot2_3.4.2      magrittr_2.0.3     maptools_1.1-8
[43] scales_1.2.1       codetools_0.2-18   units_0.8-0
[46] colorspace_2.1-0   utf8_1.2.3         KernSmooth_2.23-20
[49] proxy_0.4-27       stringi_1.7.12     wk_0.7.0
[52] munsell_0.5.0      zoo_1.8-11
```

### Environment

> packageVersion("StormR")
[1] ‘0.1.0’
