### 2023-08-21 #5766 @highamm
Hi @thomasarsouze,

Would you be able to add the code generating the images in Figure 2 in <https://github.com/umr-amap/StormR/tree/master/paper/images> (or, if I've missed the code generating that Figure, could you point me in the right direction?)

Thanks!
Matt
