### 2023-08-21 #5766 @highamm
Hi @thomasarsouze,

I've opened up an issue at <https://github.com/umr-amap/StormR/issues/68>.

Thanks, and hope all is well!
Matt
