### 2023-08-31 umr-amap/StormR#68 @thomasarsouze
Hi @highamm. Sorry for the delay in answering your issue, but I've had some days off. Anyway...
Neither myself or @thomaspibanez are able to reproduce this behaviour, even though I have a similar system to yours. I have a higher version of `sf_1.0-14` which might help here. Otherwise please try update other libraries too.
