### 2023-08-31 #5766 @thomasarsouze
> Would you be able to add the code generating the images in Figure 2 in https://github.com/umr-amap/StormR/tree/master/paper/images (or, if I've missed the code generating that Figure, could you point me in the right direction?)

As stated in the caption of the figure, all plots were generated using `plotStorms` and `plotBehaviour` functions, and the code to produce them can be seen on [this `Spatial Behaviour` article](https://umr-amap.github.io/StormR/articles/SpatialBehaviour.html). However, they were gathered into one figure manualy.
