### 2023-08-31 umr-amap/StormR#68 @highamm
Hi @thomasarsouze,

I had to install the optional `rworldxtra` package to fix the issue. Additionally, I have verified that if I remove the `rworldextra` package with `remove.packages("rworldextra")`, restart `R`, and try to run the `"Boose"` method code again, then the error reappears:

```
remove.packages("rworldxtra")
## restart R
library(StormR)
sds <- defStormsDataset()
st <- defStormsList(sds = sds, loi = c(168.33, -17.73),
                    names = "PAM", verbose = 0)

pf <- spatialBehaviour(st, product = "Profiles",
                       method = "Boose", verbose = 0)
```

Are you able to reproduce this behavior? Do users always need `rworldextra` to use the `"Boose"` method or is this more an issue with this particular data set on my particular machine?

Thanks!
