### 2023-08-31 #5766 @highamm
Hey @thomasarsouze, thanks for the info! Would it be possible to add a link to the Spatial Behaviour article you reference in the Figure caption itself so that people reading the article know where to go to reproduce the sub-figures easily?
