### 2023-08-31 umr-amap/StormR#68 @thomasarsouze
Indeed, I can reproduce the error when  removing `rworldxtra`. However, the error message is different from the one above and explicitely states that the package is missing.

This is because the `getMap` function from `rworldmap` requires `rworldxtra` when using high resolution (which is the case for the Boose method). I would have expected a warning message instead of an error.
