### 2023-08-31 f25bf7b
Update paper.md: add reference to reproduce plots
