+++ pkgdown.yml
@@ -8,5 +8,5 @@ articles:
   Models: Models.html
   SpatialBehaviour: SpatialBehaviour.html
   TemporalBehaviour: TemporalBehaviour.html
-last_built: 2023-08-31T12:13Z
+last_built: 2023-08-31T12:39Z
 
