### 2023-09-01 #5766 @highamm
Hello @thomasarsouze, can you add some community guidelines (clear guidelines for third parties wishing to 1. Contribute to the software, 2. Report issues or problems with the software, 3. Seek support) to the `StormR` repository README file?

Or, if these are given elsewhere, could you point me in the direction of where they are given? Thanks!
