### 2023-09-04 38cb634
Add community guidelines to README
