### 2023-09-04 #5766 @thomasarsouze
>can you add some community guidelines (clear guidelines for third parties wishing to 1. Contribute to the software, 2. Report issues or problems with the software, 3. Seek support) to the StormR repository README file?

Indeed this is missing information.
Done in [commit 38cb634](https://github.com/umr-amap/StormR/commit/38cb6342d1ccb5bd9eff2ef0855309a30ba5a751)
