+++ R/writeRast.R
@@ -1,8 +1,3 @@
-
-
-
-
-
 #' Check inputs for writeRast function
 #'
 #' @noRd
@@ -54,7 +49,7 @@ writeNC <- function(rastds, filename) {
     storm <- strsplit(lyr, split = "_")[[1]][1]
     if (product == "Speed") {
       zdimt[[j]] <- ncdf4::ncdim_def(paste0("time_", storm), "seconds since 1970-1-1 00:00:00",
-        rastds[lyr]@ptr$time,
+        rastds[lyr]@pnt$time,
         unlim = TRUE
       )
       j <- j + 1
@@ -106,7 +101,7 @@ writeNC <- function(rastds, filename) {
     ncdf4::ncatt_put(ncobj, ncvars[[n + 1]], "crs_wkt", prj, prec = "text")
     # need for older gdal?
     ncdf4::ncatt_put(ncobj, ncvars[[n + 1]], "spatial_ref", prj, prec = "text")
-    prj <- rastds[1]@ptr$get_crs("proj4")
+    prj <- rastds[1]@pnt$get_crs("proj4")
     if (prj != "") {
       ncdf4::ncatt_put(ncobj, ncvars[[n + 1]], "proj4", prj, prec = "text")
     }
@@ -251,8 +246,8 @@ writeRast <- function(rast, filename = NULL, path = "./", ...) {
 
     # We recreate a spatRasterDataset, one subDataset is one product for one storm
     splitVector <- c()
-    for (i in 1:rast@ptr$nlyr()) {
-      ind <- which(paste(c(strsplit(rast[[i]]@ptr$names, split = "_", fixed = TRUE)[[1]][1:2]),
+    for (i in 1:rast@pnt$nlyr()) {
+      ind <- which(paste(c(strsplit(rast[[i]]@pnt$names, split = "_", fixed = TRUE)[[1]][1:2]),
         collapse = "_"
       ) == varNames)
       splitVector <- c(splitVector, ind)
