+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: StormR
 Title: R Package to Compute the Behaviour of Wind Generated by Tropical
     Storms and Cyclones
 Maintainer: Thomas Arsouze <thomas.arsouze@cirad.fr>
-Version: 0.1.0
+Version: 0.1.1
 URL: https://umr-amap.github.io/StormR/
 BugReports: https://github.com/umr-amap/StormR/issues/new/choose
 Authors@R: c(
