+++ pkgdown.yml
@@ -8,5 +8,5 @@ articles:
   Models: Models.html
   SpatialBehaviour: SpatialBehaviour.html
   TemporalBehaviour: TemporalBehaviour.html
-last_built: 2023-10-06T11:05Z
+last_built: 2023-10-06T12:13Z
 
