### 2023-10-31 0a91bcc
Bugfix: temporalBehaviour with TS product and >2 points. Solves #71.
