### 2023-11-01 47beceb
Update SpatRaster accessors
