### 2023-11-07 #5766 @hugoledoux
Thanks @gsapijaszko for coming back to us.

You only mention the "paper", but JOSS is mostly about the code, the docs, and all the criteria as listed above. Can you let us know whether you installed the software and tested all the criteria? I am a bit confused by your wording here.
