### 2023-11-22 7a94fb5
Adapt code de CRAN comments
