+++ R/defStormsList.R
@@ -41,6 +41,14 @@
 #'   \item `pres`, Pressure at the center (pa)
 #'   \item `poci`, Pressure of the Outermost Closed Isobar (pa)
 #' }
+#' @return A `storm` object.
+#' \itemize{
+#'  \item `name`, character.
+#'  \item `season`, numeric.
+#'  \item `sshs`, numeric.
+#'  \item `obs`, numeric.
+#'  \item `obs.all`, data.frame.
+#' }
 #' @importFrom methods new
 #' @export
 storm <- methods::setClass("storm",
@@ -73,6 +81,13 @@ setOldClass("sf")
 #' @slot spatialLoi sf object. Represents the location of interest. Projection
 #'   is EPSG:4326
 #' @slot spatialLoiBuffer sf object. Buffer extension of `spatialLoi`
+#' @return A `stormsList` object.
+#' \itemize{
+#'  \item `data`, list.
+#'  \item `buffer`, numeric.
+#'  \item `spatialLoi`, sf.
+#'  \item `spatialLoiBuffer`, sf.
+#' }
 #' @importFrom methods new
 #' @importFrom methods show
 #' @export
@@ -98,15 +113,17 @@ stormsList <- methods::setClass("stormsList",
 #'
 #' Display the `storm`/`stormsList` object
 #'
-#' @noRd
 #' @param object `storm`/`stormsList` object
 #'
-#' @return NULL
+#' @return NULL, only display information about the object
+#'
+#' @export
 #' @docType methods
 #' @rdname show-methods
 #' @examples
-#' \dontrun{
-#' sts <- storms(loi = "New Caledonia", names = c("ERICA","NIRAN"))
+#' \donttest{
+#' sds <- defStormsDataset()
+#' sts <- defStormsList(sds = sds, loi = "New Caledonia", names = c("LUCAS","NIRAN"))
 #' ## Display information about Niran in sts
 #' getStorm(sts, name = "NIRAN")
 #'
@@ -164,17 +181,17 @@ setMethod("show",
 #' years should be specified, with cyclones assigned for the year
 #' #'  that originated in. By default all storms occurring since
 #' 1980 are extracted.
-#' @return The `getStorm()` function returns a `storm` object.
+#' @return A `storm` object.
 #' @export
 #' @docType methods
 #' @rdname getStorm-methods
 #' @examples
-#' \dontrun{
 #' #Creating a stormsDataset
+#' \donttest{
 #' sds <- defStormsDataset()
 #'
 #' #Getting storm track data for all storms near New Caledonia
-#' sts <- storms(sds=sds, loi = "New Caledonia")
+#' sts <- defStormsList(sds=sds, loi = "New Caledonia")
 #'
 #' #Getting `storm` for the tropical cyclone Niran
 #' st <- getStorm(sts, name = "NIRAN")
@@ -208,17 +225,17 @@ setMethod("getStorm", signature("stormsList"), function(sts, name, season = NULL
 #'
 #' @param sts `stormsList`
 #'
-#' @return numeric.
+#' @return numeric, the number of `storm` objects.
 #' @export
 #' @docType methods
 #' @rdname getNbStorms-methods
 #' @examples
-#' \dontrun{
 #' #Creating a stormsDataset
+#' \donttest{
 #' sds <- defStormsDataset()
 #'
 #' #Getting storm track data for all storms near New Caledonia
-#' sts <- storms(sds=sds, loi = "New Caledonia")
+#' sts <- defStormsList(sds=sds, loi = "New Caledonia")
 #'
 #' #Getting the number of storms in the sts object
 #' getNbStorms(sts)
[…]
