+++ vignettes/Models.Rmd
@@ -197,11 +197,15 @@ $\phi$ is the latitude. <br />
 Here, we compare wind fields generated by different models that can be used in
 `StormR` for the same time and location (tropical cyclone Pam near Vanuatu)
 
+``` {r , include = FALSE}
+oldpar <- par(mfrow = c(1,2))
+```
+
 ```{r}
 sds <- defStormsDataset()
 st <- defStormsList(sds = sds, loi = c(168.33, -17.73), names = "PAM", verbose = 0)
 PAM <- getObs(st, name = "PAM")
-par(mfrow = c(1, 2))
+
 pf <- spatialBehaviour(st, product = "Profiles", method = "Holland", asymmetry = "None", verbose = 0)
 terra::plot(pf$PAM_Speed_41, main = "Holland (1980)", cex.main = 0.8, range = c(0, 90))
 terra::plot(countriesHigh, add = TRUE)
@@ -210,7 +214,7 @@ pf <- spatialBehaviour(st, product = "Profiles", method = "Willoughby", asymmetr
 terra::plot(pf$PAM_Speed_41, main = "Willoughby et al. (2006)", cex.main = 0.8, range = c(0, 90))
 terra::plot(countriesHigh, add = TRUE)
 lines(PAM$lon, PAM$lat, lty = 3)
-par(mfrow = c(1, 2))
+
 pf <- spatialBehaviour(st, product = "Profiles", method = "Holland", asymmetry = "Miyazaki", verbose = 0)
 terra::plot(pf$PAM_Speed_41, main = "Holland (1980) + Miyazaki et al. (1962)", cex.main = 0.8, range = c(0, 90))
 terra::plot(countriesHigh, add = TRUE)
@@ -219,7 +223,7 @@ pf <- spatialBehaviour(st, product = "Profiles", method = "Willoughby", asymmetr
 terra::plot(pf$PAM_Speed_41, main = "Willoughby et al. (2006) + Miyazaki et al. (1962)", cex.main = 0.8, range = c(0, 90))
 terra::plot(countriesHigh, add = TRUE)
 lines(PAM$lon, PAM$lat, lty = 3)
-par(mfrow = c(1, 2))
+
 pf <- spatialBehaviour(st, product = "Profiles", method = "Holland", asymmetry = "Chen", verbose = 0)
 terra::plot(pf$PAM_Speed_41, main = "Holland (1980) + Chen (1994)", cex.main = 0.8, range = c(0, 90))
 terra::plot(countriesHigh, add = TRUE)
@@ -228,15 +232,17 @@ pf <- spatialBehaviour(st, product = "Profiles", method = "Willoughby", asymmetr
 terra::plot(pf$PAM_Speed_41, main = "Willoughby et al. (2006) + Chen (1994)", cex.main = 0.8, range = c(0, 90))
 terra::plot(countriesHigh, add = TRUE)
 lines(PAM$lon, PAM$lat, lty = 3)
-par(mfrow = c(1, 2))
+
 pf <- spatialBehaviour(st, product = "Profiles", method = "Boose", verbose = 0)
 terra::plot(pf$PAM_Speed_41, main = "Boose et al. (2004)", cex.main = 0.8, range = c(0, 90))
 terra::plot(countriesHigh, add = TRUE)
 lines(PAM$lon, PAM$lat, lty = 3)
+par(oldpar)
 ```
 
-``` {r ending, include = FALSE}
-par()
+``` {r , include = FALSE}
+par(oldpar)
 ```
+
 ## References
 
