+++ DESCRIPTION
@@ -20,7 +20,7 @@ Authors@R: c(
   )
 Description: Set of functions to quantify and map the
     behaviour of winds generated by tropical storms and cyclones in space
-    and time. It includes functions to computes and analyze fields such as the maximum sustained wind field, power dissipation index and duration of exposure to winds above a given threshold. It also includes functions to map the trajectories as well as characteristics of the storms.
+    and time. It includes functions to compute and analyze fields such as the maximum sustained wind field, power dissipation index and duration of exposure to winds above a given threshold. It also includes functions to map the trajectories as well as characteristics of the storms.
 License: GPL (>= 3)
 Depends: 
     R (>= 2.10)
