+++ man/show-methods.Rd
@@ -30,3 +30,4 @@ getStorm(sts, name = "NIRAN")
 sts
 }
 }
+\keyword{internal}
