### 2023-11-23 e3c9fc9
Bugfix pkgdown build
