### 2023-11-24 f86f315
Update README with CRAN install
