### 2023-11-24 1201832
Add badges + README typos
