### 2023-11-27 1a6d2f1
Add plotTemporal function
