### 2023-11-27 e141ff2
Changes in observation intervals for behaviour functions
