+++ R/temporalBehaviour.R
@@ -17,7 +17,7 @@ checkInputsTemporalBehaviour <- function(sts, points, product, windThreshold, me
                                          empiricalRMW, tempRes, verbose) {
   # Checking sts input
   stopifnot("no data found" = !missing(sts))
-
+  
   # Checking points input
   stopifnot("no data found" = !missing(points))
   stopifnot("points must be data.frame" = identical(class(points), "data.frame"))
@@ -25,34 +25,34 @@ checkInputsTemporalBehaviour <- function(sts, points, product, windThreshold, me
     "colnames of points must be \"x\" (Eastern degree), \"y\" (Northern degree)" = colnames(points) == c("x", "y")
   )
   stopifnot("Invalid points coordinates" = points$x > -180 & points$x <= 360 &
-    points$y >= -90 & points$y <= 90)
-
+              points$y >= -90 & points$y <= 90)
+  
   # Checking product input
   stopifnot("Invalid product" = product %in% c("TS", "PDI", "Exposure"))
   stopifnot("Only one product must be chosen" = length(product) == 1)
-
+  
   # Checking windThreshold input
   if ("Exposure" %in% product) {
     stopifnot("windThreshold must be numeric" = identical(class(windThreshold), "numeric"))
     stopifnot("invalid value(s) for windThreshold input (must be > 0)" = windThreshold > 0)
   }
-
+  
   # Checking method input
   stopifnot("Invalid method input" = method %in% c("Willoughby", "Holland", "Boose"))
   stopifnot("Only one method must be chosen" = length(method) == 1)
-
+  
   # Checking asymmetry input
   stopifnot("Invalid asymmetry input" = asymmetry %in% c("None", "Chen", "Miyazaki"))
   stopifnot("Only one asymmetry must be chosen" = length(asymmetry) == 1)
-
+  
   # Checking empiricalRMW input
   stopifnot("empiricalRMW must be logical" = identical(class(empiricalRMW), "logical"))
-
+  
   # Checking tempRes input
   stopifnot("tempRes must be numeric" = identical(class(tempRes), "numeric"))
   stopifnot("tempRes must be length 1" = length(tempRes) == 1)
-  stopifnot("invalid tempRes: must be either 1, 0.75, 0.5 or 0.25" = tempRes %in% c(1, 0.75, 0.5, 0.25))
-
+  stopifnot("invalid tempRes: must be either 60, 30 or 15" = tempRes %in% c(60, 30, 15))
+  
   # Checking verbose input
   stopifnot("verbose must be numeric" = identical(class(verbose), "numeric"))
   stopifnot("verbose must length 1" = length(verbose) == 1)
@@ -81,7 +81,7 @@ computePDI <- function(wind, tempRes) {
   pdi <- pdi * rho * cd
   # Integrating over the whole track
   pdi <- sum(pdi, na.rm = TRUE) * tempRes
-
+  
   return(round(pdi, 3))
 }
 
@@ -107,7 +107,7 @@ computeExposure <- function(wind, tempRes, threshold) {
     expo[ind] <- 1
     exposure <- c(exposure, sum(expo, na.rm = TRUE) * tempRes)
   }
-
+  
   return(exposure)
 }
 
@@ -141,7 +141,7 @@ computeProduct <- function(product, wind, direction, tempRes, result, threshold)
   } else if (product == "Exposure") {
     prod <- computeExposure(wind, tempRes, threshold)
   }
-
+  
   return(cbind(result, prod))
 }
 
@@ -167,10 +167,10 @@ finalizeResult <- function(finalResult, result, product, points, isoT, indices, 
     l <- list()
     for (i in 1:dim(points)[1]) {
       j <- 2 * (i - 1) + 1
-
+      
[…]
