+++ tests/testthat/test-spatialBehaviour.R
@@ -30,7 +30,7 @@ test_that("Test Willougbhy / Holland / Boose functions", {
       28.131, 27.988, 27.846, 27.706, 27.567, 27.430, 27.294, 27.159, 27.026, 26.895
     )
   )
-
+  
   expect_equal(
     holland(seq(1, 100), 20, 50, 1020 * 100, 915 * 100, -15),
     c(
@@ -46,7 +46,7 @@ test_that("Test Willougbhy / Holland / Boose functions", {
       30.931, 30.667, 30.406, 30.147, 29.890, 29.635, 29.381, 29.130, 28.881, 28.633
     )
   )
-
+  
   expect_equal(
     boose(seq(1, 100), 20, 50, 1020 * 100, 915 * 100, 0.1, 0.1, 0.25, 0.25, 15, 1, -15),
     c(
@@ -68,7 +68,7 @@ test_that("Test spatialBehaviour function", {
   suppressWarnings(sds <- defStormsDataset(verbose = 0))
   pam <- defStormsList(sds = sds, loi = "Vanuatu", names = "PAM", verbose = 0)
   sb <- spatialBehaviour(pam, method = "Holland", verbose = 0)
-
+  
   expect_equal(
     terra::minmax(sb[["PAM_MSW"]]),
     matrix(
@@ -76,7 +76,7 @@ test_that("Test spatialBehaviour function", {
       dimnames =list(c('min', 'max'), c('PAM_MSW'))
     )
   )
-
+  
   sb <- spatialBehaviour(pam, method = "Boose", verbose = 0)
   expect_equal(
     terra::minmax(sb[["PAM_MSW"]]),
@@ -98,10 +98,10 @@ test_that("Test checkInputsSpatialBehaviour function", {
     asymmetry = "Chen",
     empiricalRMW = FALSE,
     spaceRes = "2.5min",
-    tempRes = 1,
+    tempRes = 60,
     verbose = 2
   ))
-
+  
   # Checking product input
   expect_error(checkInputsSpatialBehaviour(
     sts = pam,
@@ -111,10 +111,10 @@ test_that("Test checkInputsSpatialBehaviour function", {
     asymmetry = "Chen",
     empiricalRMW = FALSE,
     spaceRes = "2.5min",
-    tempRes = 1,
+    tempRes = 60,
     verbose = 2
   ))
-
+  
   expect_error(checkInputsSpatialBehaviour(
     sts = pam,
     product = "TS",
@@ -123,10 +123,10 @@ test_that("Test checkInputsSpatialBehaviour function", {
     asymmetry = "Chen",
     empiricalRMW = FALSE,
     spaceRes = "2.5min",
-    tempRes = 1,
+    tempRes = 60,
     verbose = 2
   ))
-
+  
   # Check windThreshold input
   expect_error(checkInputsSpatialBehaviour(
     sts = pam,
@@ -136,10 +136,10 @@ test_that("Test checkInputsSpatialBehaviour function", {
     asymmetry = "Chen",
     empiricalRMW = FALSE,
     spaceRes = "2.5min",
-    tempRes = 1,
+    tempRes = 60,
     verbose = 2
   ))
-
+  
   expect_error(checkInputsSpatialBehaviour(
     sts = pam,
     product = c("MSW", "PDI", "Exposure"),
@@ -148,10 +148,10 @@ test_that("Test checkInputsSpatialBehaviour function", {
     asymmetry = "Chen",
     empiricalRMW = FALSE,
     spaceRes = "2.5min",
-    tempRes = 1,
+    tempRes = 60,
     verbose = 2
   ))
-
+  
   # Checking method input
   expect_error(checkInputsSpatialBehaviour(
     sts = pam,
[…]

[…]
