+++ man/plotTemporal.Rd
@@ -1,0 +1,32 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/plotTemporal.R
+\name{plotTemporal}
+\alias{plotTemporal}
+\title{Plotting wind behaviour time series and summary statistics at given point
+locations}
+\usage{
+plotTemporal(data, storm, var = "speed")
+}
+\arguments{
+\item{data}{time series generated with \code{temporalBehaviour} with \code{product=TS}
+input}
+
+\item{storm}{list of characters. Storm names. The storm must be available in
+\code{data} input. It can also be a vector of integer corresponding to the
+indices of storms stored in \code{data}input.}
+
+\item{var}{character. Represent the type of variable to plot. Must be either
+\code{speed} or \code{direction}. Default value is set to \code{speed}.}
+}
+\value{
+null
+}
+\description{
+Plotting wind behaviour time series and summary statistics at given point
+locations
+}
+\examples{
+\donttest{
+#Add example here
+}
+}
