+++ NAMESPACE
@@ -15,6 +15,7 @@ export(getSeasons)
 export(getStorm)
 export(plotBehaviour)
 export(plotStorms)
+export(plotTemporal)
 export(spatialBehaviour)
 export(storm)
 export(stormsDataset)
