### 2023-11-28 002d2f2
fix few errors/warnings/notes - 2 remaining
