### 2023-11-28 68e24ae
replace sshs input with scale input to be more generic
