+++ R/defStormsDataset.R
@@ -3,9 +3,9 @@
 
 
 
-###########################
+############################
 # Unit conversion functions#
-###########################
+############################
 
 # For msw
 knt2ms <- function(x) {
@@ -72,7 +72,7 @@ atm2pa <- function(x) {
 #' is either a 1D array of dimension (number of storms) for `names` and
 #' `seasons` fields, or a 2D array of dimension
 #' (Maximum number of observations:number of storms), for the remaining fields
-#' which are `isoTime`, `lon`, `lat`, `msw`, `rmw`, `pressure`, `poci`, `sshs`
+#' which are `isoTime`, `lon`, `lat`, `msw`, `rmw`, `pressure`, `poci`, `scale`
 #'
 #' @details
 #' The fields input must provide at least 6 mandatory fields (and at most 11) in
@@ -102,9 +102,8 @@ atm2pa <- function(x) {
 #'  \item A field `rmw`: which dimension contains the radius of maximum
 #'        wind speed of each observations for all storms in the netcdf
 #'        database (See spatialBehaviour, temporalBehaviour)
-#'  \item A field `sshs`: which dimension contains the Saffir Simpson
-#'        Hurricane Scale index of each observations for all storms in the
-#'        netcdf database
+#'  \item A field `scale`: which dimension contains the storm scale index of
+#'        each observations for all storms in the netcdf database
 #' }
 #' Finally these following fields are optional but mandatory to perform Holland
 #' model (See `spatialBehaviour`, `temporalBehaviour`)
@@ -121,7 +120,7 @@ atm2pa <- function(x) {
 #' databases:
 #' `fields = c(basin = "basin", names = "name", seasons = "season", isoTime = "iso_time",
 #' lon = "usa_lon", lat = "usa_lat", msw = "usa_wind", rmw = "usa_rmw", pressure = "usa_pres",
-#' poci = "usa_poci", sshs = "usa_sshs")`
+#' poci = "usa_poci", scale = "usa_sshs")`
 #'
 #' @export
 stormsDataset <- methods::setClass(
@@ -268,7 +267,7 @@ checkInputsdefStormsDataset <- function(filename, fields, basin, seasons, unitCo
 #'        its band of strongest winds (recommended),
 #'    \item `"pressure"`, central pressure (recommended),
 #'    \item `"poci"`, pressure of the last closed isobar (recommended), and
-#'    \item `"sshs"`, Saffir-Simpson hurricane wind scale rating based on msw (optional).
+#'    \item `"scale"`, Storm scale rating (optional).
 #'  }
 #' @param basin character. If the basin field is provided, then storm track data will
 #' only be extracted for the named basin. By default `basin=NULL`, meaning that all storms
@@ -340,7 +339,7 @@ defStormsDataset <- function(filename = system.file("extdata", "test_dataset.nc"
                                lat = "usa_lat",
                                msw = "usa_wind",
                                basin = "basin",
-                               sshs = "usa_sshs",
+                               scale = "usa_sshs",
                                rmw = "usa_rmw",
                                pressure = "usa_pres",
                                poci = "usa_poci"
@@ -433,9 +432,9 @@ defStormsDataset <- function(filename = system.file("extdata", "test_dataset.nc"
   data$latitude <- data$latitude[, o]
   data$msw <- data$msw[, o]
 
-  if ("sshs" %in% names(fields)) {
-    data$sshs <- array(ncdf4::ncvar_get(dataBase, fields["sshs"])[, ind], dim = c(row, len))
-    data$sshs <- data$sshs[, o]
+  if ("scale" %in% names(fields)) {
[…]
