+++ tests/testthat/test-stormsDataset.R
@@ -2,902 +2,1104 @@
 
 
 
+
 test_that("Test checkInputsdefStormsDataset function", {
   # Checking filename input
-  expect_error(checkInputsdefStormsDataset(
-    fields = c(
-      "basin" = "basin",
-      "names" = "name",
-      "seasons" = "season",
-      "isoTime" = "iso_time",
-      "lon" = "usa_lon",
-      "lat" = "usa_lat",
-      "msw" = "usa_wind",
-      "sshs" = "usa_sshs",
-      "rmw" = "usa_rmw",
-      "pressure" = "usa_pres",
-      "poci" = "usa_poci"
-    ),
-    basin = "SP",
-    seasons = c(1980, as.numeric(format(Sys.time(), "%Y"))),
-    unitConversion = c(
-      msw = "knt2ms",
-      rmw = "nm2km",
-      pressure = "mb2pa",
-      poci = "mb2pa"
-    ),
-    verbose = 1
-  ))
-
-  expect_error(checkInputsdefStormsDataset(
-    filename = 1,
-    fields = c(
-      "basin" = "basin",
-      "names" = "name",
-      "seasons" = "season",
-      "isoTime" = "iso_time",
-      "lon" = "usa_lon",
-      "lat" = "usa_lat",
-      "msw" = "usa_wind",
-      "sshs" = "usa_sshs",
-      "rmw" = "usa_rmw",
-      "pressure" = "usa_pres",
-      "poci" = "usa_poci"
-    ),
-    basin = "SP",
-    seasons = c(1980, as.numeric(format(Sys.time(), "%Y"))),
-    unitConversion = c(
-      msw = "knt2ms",
-      rmw = "nm2km",
-      pressure = "mb2pa",
-      poci = "mb2pa"
-    ),
-    verbose = 1
-  ))
-
-  expect_error(checkInputsdefStormsDataset(
-    filename = TRUE,
-    fields = c(
-      "basin" = "basin",
-      "names" = "name",
-      "seasons" = "season",
-      "isoTime" = "iso_time",
-      "lon" = "usa_lon",
-      "lat" = "usa_lat",
-      "msw" = "usa_wind",
-      "sshs" = "usa_sshs",
-      "rmw" = "usa_rmw",
-      "pressure" = "usa_pres",
-      "poci" = "usa_poci"
-    ),
-    basin = "SP",
-    seasons = c(1980, as.numeric(format(Sys.time(), "%Y"))),
-    unitConversion = c(
-      msw = "knt2ms",
-      rmw = "nm2km",
-      pressure = "mb2pa",
-      poci = "mb2pa"
-    ),
-    verbose = 1
-  ))
-
-
-  expect_error(checkInputsdefStormsDataset(
-    filename = c("data", ".nc"),
-    fields = c(
-      "basin" = "basin",
-      "names" = "name",
-      "seasons" = "season",
-      "isoTime" = "iso_time",
-      "lon" = "usa_lon",
-      "lat" = "usa_lat",
-      "msw" = "usa_wind",
-      "sshs" = "usa_sshs",
-      "rmw" = "usa_rmw",
-      "pressure" = "usa_pres",
-      "poci" = "usa_poci"
-    ),
-    basin = "SP",
-    seasons = c(1980, as.numeric(format(Sys.time(), "%Y"))),
-    unitConversion = c(
-      msw = "knt2ms",
-      rmw = "nm2km",
-      pressure = "mb2pa",
-      poci = "mb2pa"
-    ),
-    verbose = 1
-  ))
-
+  expect_error(
+    checkInputsdefStormsDataset(
+      fields = c(
+        "basin" = "basin",
+        "names" = "name",
+        "seasons" = "season",
+        "isoTime" = "iso_time",
[…]
