### 2023-11-29 fec913a
add scale_Palette input in stormDataset
