+++ tests/testthat/test-stormsDataset.R
@@ -3,6 +3,7 @@
 
 
 
+
 test_that("Test checkInputsdefStormsDataset function", {
   # Checking filename input
   expect_error(
@@ -22,6 +23,7 @@ test_that("Test checkInputsdefStormsDataset function", {
       ),
       basin = "SP",
       seasons = c(1980, as.numeric(format(Sys.time(
+        
       ), "%Y"))),
       unitConversion = c(
         msw = "knt2ms",
@@ -30,6 +32,7 @@ test_that("Test checkInputsdefStormsDataset function", {
         poci = "mb2pa"
       ),
       scale = sshs,
+      scalePalette = SSHS_PALETTE,
       verbose = 1
     )
   )
@@ -52,6 +55,7 @@ test_that("Test checkInputsdefStormsDataset function", {
       ),
       basin = "SP",
       seasons = c(1980, as.numeric(format(Sys.time(
+        
       ), "%Y"))),
       unitConversion = c(
         msw = "knt2ms",
@@ -60,6 +64,7 @@ test_that("Test checkInputsdefStormsDataset function", {
         poci = "mb2pa"
       ),
       scale = sshs,
+      scalePalette = SSHS_PALETTE,
       verbose = 1
     )
   )
@@ -82,6 +87,7 @@ test_that("Test checkInputsdefStormsDataset function", {
       ),
       basin = "SP",
       seasons = c(1980, as.numeric(format(Sys.time(
+        
       ), "%Y"))),
       unitConversion = c(
         msw = "knt2ms",
@@ -90,6 +96,7 @@ test_that("Test checkInputsdefStormsDataset function", {
         poci = "mb2pa"
       ),
       scale = sshs,
+      scalePalette = SSHS_PALETTE,
       verbose = 1
     )
   )
@@ -113,6 +120,7 @@ test_that("Test checkInputsdefStormsDataset function", {
       ),
       basin = "SP",
       seasons = c(1980, as.numeric(format(Sys.time(
+        
       ), "%Y"))),
       unitConversion = c(
         msw = "knt2ms",
@@ -121,6 +129,7 @@ test_that("Test checkInputsdefStormsDataset function", {
         poci = "mb2pa"
       ),
       scale = sshs,
+      scalePalette = SSHS_PALETTE,
       verbose = 1
     )
   )
@@ -132,6 +141,7 @@ test_that("Test checkInputsdefStormsDataset function", {
       fields = 1,
       basin = "SP",
       seasons = c(1980, as.numeric(format(Sys.time(
+        
       ), "%Y"))),
       unitConversion = c(
         msw = "knt2ms",
@@ -140,6 +150,7 @@ test_that("Test checkInputsdefStormsDataset function", {
         poci = "mb2pa"
       ),
       scale = sshs,
+      scalePalette = SSHS_PALETTE,
       verbose = 1
     )
   )
@@ -150,6 +161,7 @@ test_that("Test checkInputsdefStormsDataset function", {
       fields = TRUE,
       basin = "SP",
       seasons = c(1980, as.numeric(format(Sys.time(
+        
       ), "%Y"))),
       unitConversion = c(
         msw = "knt2ms",
@@ -158,6 +170,7 @@ test_that("Test checkInputsdefStormsDataset function", {
         poci = "mb2pa"
       ),
       scale = sshs,
+      scalePalette = SSHS_PALETTE,
       verbose = 1
     )
   )
@@ -179,6 +192,7 @@ test_that("Test checkInputsdefStormsDataset function", {
       ),
       basin = "SP",
       seasons = c(1980, as.numeric(format(Sys.time(
+        
       ), "%Y"))),
       unitConversion = c(
         msw = "knt2ms",
@@ -187,6 +201,7 @@ test_that("Test checkInputsdefStormsDataset function", {
[…]
