### 2023-11-30 4872556
Remove all remaining hardcoded sshs input replace with more general scale input
