+++ tests/testthat/test-spatialBehaviour.R
@@ -3,11 +3,11 @@
 
 
 
+
 test_that("Test getRmw function", {
-  expect_equal(getRmw(seq(0, 80, 5), -15), c(
-    60, 55, 51, 47, 44, 41, 38, 35, 32,
-    30, 28, 25, 24, 22, 20, 19, 17
-  ))
+  expect_equal(getRmw(seq(0, 80, 5),-15),
+               c(60, 55, 51, 47, 44, 41, 38, 35, 32,
+                 30, 28, 25, 24, 22, 20, 19, 17))
 })
 
 
@@ -16,301 +16,650 @@ test_that("Test getRmw function", {
 
 test_that("Test Willougbhy / Holland / Boose functions", {
   expect_equal(
-    willoughby(seq(1, 100), 20, 50, -15),
+    willoughby(seq(1, 100), 20, 50,-15),
     c(
-      1.813, 3.905, 6.118, 8.414, 10.772, 13.182, 15.635, 18.127, 20.652, 23.208,
-      25.791, 28.399, 31.032, 33.685, 36.360, 39.053, 41.765, 44.494, 47.239, 50.000,
-      49.327, 48.676, 48.044, 47.433, 46.839, 46.264, 45.706, 45.165, 44.640, 44.131,
-      43.636, 43.155, 42.688, 42.235, 41.794, 41.366, 40.949, 40.543, 40.149, 39.765,
-      39.391, 39.027, 38.672, 38.327, 37.990, 37.661, 37.341, 37.028, 36.723, 36.425,
-      36.134, 35.850, 35.572, 35.300, 35.035, 34.775, 34.521, 34.272, 34.028, 33.789,
-      33.555, 33.326, 33.102, 32.881, 32.665, 32.453, 32.245, 32.040, 31.839, 31.642,
-      31.448, 31.257, 31.070, 30.885, 30.704, 30.525, 30.349, 30.176, 30.005, 29.837,
-      29.671, 29.508, 29.347, 29.188, 29.031, 28.876, 28.724, 28.573, 28.424, 28.277,
-      28.131, 27.988, 27.846, 27.706, 27.567, 27.430, 27.294, 27.159, 27.026, 26.895
+      1.813,
+      3.905,
+      6.118,
+      8.414,
+      10.772,
+      13.182,
+      15.635,
+      18.127,
+      20.652,
+      23.208,
+      25.791,
+      28.399,
+      31.032,
+      33.685,
+      36.360,
+      39.053,
+      41.765,
+      44.494,
+      47.239,
+      50.000,
+      49.327,
+      48.676,
+      48.044,
+      47.433,
+      46.839,
+      46.264,
+      45.706,
+      45.165,
+      44.640,
+      44.131,
+      43.636,
+      43.155,
+      42.688,
+      42.235,
+      41.794,
+      41.366,
+      40.949,
+      40.543,
+      40.149,
+      39.765,
+      39.391,
+      39.027,
+      38.672,
+      38.327,
+      37.990,
+      37.661,
+      37.341,
+      37.028,
+      36.723,
+      36.425,
+      36.134,
+      35.850,
+      35.572,
+      35.300,
+      35.035,
+      34.775,
+      34.521,
+      34.272,
+      34.028,
+      33.789,
+      33.555,
+      33.326,
+      33.102,
+      32.881,
+      32.665,
+      32.453,
+      32.245,
+      32.040,
+      31.839,
+      31.642,
+      31.448,
+      31.257,
+      31.070,
+      30.885,
+      30.704,
+      30.525,
+      30.349,
+      30.176,
+      30.005,
+      29.837,
+      29.671,
+      29.508,
+      29.347,
+      29.188,
+      29.031,
+      28.876,
+      28.724,
[…]
