+++ R/plotTemporal.R
@@ -8,24 +8,25 @@
 #' @param var character
 #' @return NULL
 checkInputsPlotTemporal <- function(data, storm, var) {
-  
+
   # Checking data input
   stopifnot("no data found" = !missing(data))
-  stopifnot("data must be temporal series. Be sure to use data generated from temporalBehaviour() function run with product input set to 'TS'" = identical(class(data[[1]]), "list"))
-  
+  stopifnot("data must be temporal series. Be sure to use data generated from temporalBehaviour() 
+  function run with product input set to 'TS'" = identical(class(data[[1]]), "list"))
+
   # Checking storm input
   stopifnot("You must select one single storm" = !missing(storm))
-  if(identical(class(storm), "character")){
+  if (identical(class(storm), "character")) {
     stopifnot("storm not find in data" = (storm %in% names(data)))
-  }else if(identical(class(storm), "numeric")){
+  }else if (identical(class(storm), "numeric")) {
     stopifnot("storm index out of bound in data" = (storm %in% c(1, length(names(data)))))
-  }else{
+  }else {
     stop("Invalid storm input")
   }
-  
+
   # Checking var input
-  stopifnot("invalid var input, must be either 'speed' or 'direction'" = 
-              var == 'speed' | var == 'direction')
+  stopifnot("invalid var input, must be either 'speed' or 'direction'" =
+              var == "speed" | var == "direction")
 
 }
 
@@ -50,119 +51,118 @@ checkInputsPlotTemporal <- function(data, storm, var) {
 #' \donttest{
 #' sds <- defStormsDataset()
 #' st <- defStormsList(sds = sds, loi = "Vanuatu", names = "PAM", verbose = 0)
-#' 
+#'
 #' df <- data.frame(x = c(168.33, 167.17), y = c(-17.73, -15.53))
 #' rownames(df) <- c("Port_Vila", "Luganville")
-#' 
+#'
 #' # Generate temporal series of wind on the points
 #' TS <- temporalBehaviour(st, points = df, product = "TS", tempRes = 0.5, verbose = 0)
-#' 
+#'
 #' # Plot temporal series of wind speed
 #' plotTemporal(data=TS, storm="PAM")
-#' 
+#'
 #' # Plot temporal series of wind direction
 #' plotTemporal(data=TS, storm="PAM", var='direction')
 #' }
-plotTemporal <- function(data, storm, var = 'speed'){
-  
+plotTemporal <- function(data, storm, var = "speed") {
+
   checkInputsPlotTemporal(data, storm, var)
-  
+
   # Filter by storm
-  sub_data <- data[storm]
-  
-  nbOfPositions <- length(sub_data[[1]])
-  
+  subData <- data[storm]
+
+  nbOfPositions <- length(subData[[1]])
+
   # Generate a sequence of colors
-  
+
   # TODO pb of dependencies here...
-  palette <- colorRampPalette(colors=c("red", "green","blue"))
+  palette <- colorRampPalette(colors = c("red", "green", "blue"))
   cols <- palette(nbOfPositions)
-  
-  # Generate a sequence of symbols 
+
+  # Generate a sequence of symbols
   symbols <- seq(1, nbOfPositions, 1)
-  
-  
-  
+
+
+
   notNaIndices <- which(!is.na(sub_data[[1]][[1]]$speed))
-                        
+
   # Define x-axis range
-  for(location in sub_data[[1]][c(2:nbOfPositions)]){
-    
+  for (location in sub_data[[1]][c(2:nbOfPositions)]) {
+
     notNaIndices <- union(notNaIndices,
[…]
