### 2023-12-04 d40ee24
new getDataFromCsvFile function
