+++ R/defStormsDataset.R
@@ -241,6 +241,292 @@ checkInputsdefStormsDataset <- function(filename, fields, basin, seasons, unitCo
 
 
 
+getDataFromNcdfFile <- function(filename, fields, basin, seasons, unitConversion, verbose){
+  
+  if (verbose) {
+    cat("=== Loading data  ===\nOpen database... ")
+  }
+  
+  
+  dataBase <- utils::read.csv(filename)
+  
+  if (verbose) {
+    cat(filename, "opened\nCollecting data ...\n")
+  }
+  
+  
+  
+  row <- dim(lon)[1] 
+  len <- dim(lon)[2] # ??
+  ind <- seq(1, len) # ??
+  
+  # Filter by season
+  ind <- which(season %in% seq(seasons[1], seasons[2], 1))
+  len <- length(ind)
+  
+  if (!is.null(basin)) {
+    # Filter by basin ID
+    basins <- ncdf4::ncvar_get(dataBase, fields["basin"])
+    indB <- which(basins[1, ] == basin)
+    ind <- intersect(ind, indB)
+    len <- length(ind)
+  }
+  
+  
+  if (unitConversion["msw"] == "mph2ms") {
+    msw <- array(mph2ms(ncdf4::ncvar_get(dataBase, fields["msw"])[, ind]),
+                 dim = c(row, len)
+    )
+  } else if (unitConversion["msw"] == "knt2ms") {
+    msw <- array(knt2ms(ncdf4::ncvar_get(dataBase, fields["msw"])[, ind]),
+                 dim = c(row, len)
+    )
+  } else if (unitConversion["msw"] == "kmh2ms") {
+    msw <- array(kmh2ms(ncdf4::ncvar_get(dataBase, fields["msw"])[, ind]),
+                 dim = c(row, len)
+    )
+  } else {
+    msw <- array(ncdf4::ncvar_get(dataBase, fields["msw"])[, ind],
+                 dim = c(row, len)
+    )
+  }
+  
+  # Collect data
+  data <- list(
+    names = ncdf4::ncvar_get(dataBase, fields["names"])[ind],
+    seasons = season[ind],
+    isotimes = array(ncdf4::ncvar_get(dataBase, fields["isoTime"])[, ind],
+                     dim = c(row, len)
+    ),
+    longitude = array(ncdf4::ncvar_get(dataBase, fields["lon"])[, ind],
+                      dim = c(row, len)
+    ),
+    latitude = array(ncdf4::ncvar_get(dataBase, fields["lat"])[, ind],
+                     dim = c(row, len)
+    ),
+    msw = msw
+  )
+  
+  # Sort by Date
+  o <- order(data$isotimes[1, ])
+  
+  data$names <- data$names[o]
+  data$seasons <- data$seasons[o]
+  data$isotimes <- data$isotimes[, o]
+  data$longitude <- data$longitude[, o]
+  data$latitude <- data$latitude[, o]
+  data$msw <- data$msw[, o]
+  
+  if ("sshs" %in% names(fields)) {
+    data$sshs <- array(ncdf4::ncvar_get(dataBase, fields["sshs"])[, ind], dim = c(row, len))
+    data$sshs <- data$sshs[, o]
+  }
+  
+  if ("rmw" %in% names(fields)) {
+    if (unitConversion["rmw"] == "nm2km") {
+      data$rmw <- array(nm2km(ncdf4::ncvar_get(dataBase, fields["rmw"])[, ind]), dim = c(row, len))
+    } else {
+      data$rmw <- array(ncdf4::ncvar_get(dataBase, fields["rmw"])[, ind], dim = c(row, len))
+    }
+    data$rmw <- data$rmw[, o]
+  }
+  
+  
[…]
