### 2023-12-05 dccc22f
bug fix in poci conversion for netcdf version
