+++ R/defStormsDataset.R
@@ -172,6 +172,10 @@ checkInputsdefStormsDataset <- function(filename, fields, basin, seasons, unitCo
     "Invalid unitConversion directive for 'msw'" =
       unitConversion["msw"] %in% c("None", "mph2ms", "knt2ms", "kmh2ms")
   )
+  
+  if(tail(strsplit(filename, "\\.")[[1]], 1) == "csv"){
+    stopifnot("No 'sid' selection in fields" = "sid" %in% names(fields))
+  }
 
   # Optional fields
   if (("basin" %in% names(fields)) && is.null(basin)) {
@@ -225,8 +229,6 @@ checkInputsdefStormsDataset <- function(filename, fields, basin, seasons, unitCo
     )
   }
 
-
-
   # Checking seasons input
   stopifnot("seasons must be numeric" = identical(class(seasons), "numeric"))
   stopifnot("seasons must be a range of calendar year" = length(seasons) == 2 & seasons[1] <= seasons[2])
@@ -247,14 +249,14 @@ getDataFromNcdfFile <- function(filename, fields, basin, seasons, unitConversion
     cat("=== Loading data  ===\nOpen database... ")
   }
   
-  
-  dataBase <- utils::read.csv(filename)
+  dataBase <- ncdf4::nc_open(filename)
   
   if (verbose) {
     cat(filename, "opened\nCollecting data ...\n")
   }
   
-  
+  lon <- ncdf4::ncvar_get(dataBase, fields["lon"])
+  season <- ncdf4::ncvar_get(dataBase, fields["seasons"])
   
   row <- dim(lon)[1] 
   len <- dim(lon)[2] # ??
@@ -347,32 +349,28 @@ getDataFromNcdfFile <- function(filename, fields, basin, seasons, unitConversion
     data$pressure <- data$pressure[, o]
   }
   
-  
   if ("poci" %in% names(fields)) {
     if (unitConversion["poci"] == "mb2pa") {
       data$poci <- array(mb2pa(ncdf4::ncvar_get(dataBase, fields["poci"])[, ind]), dim = c(row, len))
+    } else if (unitConversion["poci"] == "b2pa") {
+      data$poci <- array(b2pa(ncdf4::ncvar_get(dataBase, fields["poci"])[, ind]), dim = c(row, len))
+    } else if (unitConversion["poci"] == "psi2pa") {
+      data$poci <- array(psi2pa(ncdf4::ncvar_get(dataBase, fields["poci"])[, ind]), dim = c(row, len))
+    } else if (unitConversion["poci"] == "atm2pa") {
+      data$poci <- array(atm2pa(ncdf4::ncvar_get(dataBase, fields["poci"])[, ind]), dim = c(row, len))
     } else {
       data$poci <- array(ncdf4::ncvar_get(dataBase, fields["poci"])[, ind], dim = c(row, len))
     }
     data$poci <- data$poci[, o]
   }
   
+  
   ncdf4::nc_close(dataBase)
-  
-  
-  
-  if (verbose) {
-    cat("=== DONE ===\n")
-  }
-  
-  if (is.null(basin)) {
-    basin <- "None"
-  }
   
   return(data)
 }
 
-getDateFromCsvFile <- function(filename, fields, basin, seasons, unitConversion, verbose){
+getDataFromCsvFile <- function(filename, fields, basin, seasons, unitConversion, verbose){
   
   if (verbose) {
     cat("=== Loading data  ===\nOpen database... ")
@@ -388,10 +386,13 @@ getDateFromCsvFile <- function(filename, fields, basin, seasons, unitConversion,
   filter <- which(colnames(dataBase) %in% fields)
   dataBaseFiltered <- dataBase[, filter]
   
-
+  
+  # Remove sub header
[…]
