### 2023-12-05 33fa6a8
minor changes + documentation
