+++ man/getDataFromCsvFile.Rd
@@ -1,0 +1,29 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/defStormsDataset.R
+\name{getDataFromCsvFile}
+\alias{getDataFromCsvFile}
+\title{Load database when filename is a NetCDF
+(CF defStormsDataset for additional informations about parameters)}
+\usage{
+getDataFromCsvFile(filename, fields, basin, seasons, unitConversion, verbose)
+}
+\arguments{
+\item{filename}{character}
+
+\item{fields}{named character vecor}
+
+\item{basin}{character}
+
+\item{seasons}{numeric vector}
+
+\item{unitConversion}{named character vector}
+
+\item{verbose}{numeric}
+}
+\value{
+list of arrays
+}
+\description{
+Load database when filename is a NetCDF
+(CF defStormsDataset for additional informations about parameters)
+}
