### 2023-12-06 433389c
Minor documentation comment
