### 2023-12-06 3449a4f
Remove intermediate variables
