+++ R/defStormsDataset.R
@@ -440,19 +440,35 @@ getDataFromCsvFile <- function(filename, fields, basin, seasons, unitConversion,
   # Initialize template structure
   templateArray = array(NaN, dim=c(row,len))
   
-  # Mandatory fields
-  names <- array(NaN, dim=len)
-  seasons <- array(NaN, dim=len)
-  isotimes <- templateArray
-  longitude <- templateArray
-  latitude <- templateArray
-  msw <- templateArray
   
-  # Optional fields
-  sshs <- templateArray # TODO Remove later
-  rmw <- templateArray
-  pressure <- templateArray
-  poci <- templateArray
+  # Mandatory fields
+  data <- list(
+    names = array(NaN, dim=len),
+    seasons = array(NaN, dim=len),
+    isotimes = templateArray,
+    longitude = templateArray,
+    latitude = templateArray,
+    msw = templateArray
+  )
+  
+  # TODO Remove later
+  if("sshs" %in% names(fields)){
+    data$sshs <- templateArray
+  }
+  
+  if("rmw" %in% names(fields)){
+    data$rmw <- templateArray
+  }
+  
+  if("pressure" %in% names(fields)){
+    data$pressure <- templateArray
+  }
+  
+  if("poci" %in% names(fields)){
+    data$poci <- templateArray
+  }
+  
+
   
   for(i in 1:len){
     
@@ -465,101 +481,72 @@ getDataFromCsvFile <- function(filename, fields, basin, seasons, unitConversion,
     end = cumulativeIndex[i]
     
     # Fill data
-    
-    names[i] <- dataBaseFiltered[start, fields["names"]]
-    seasons[i] <- as.numeric(dataBaseFiltered[start, fields["seasons"]])
-    isotimes[,i] <- c(dataBaseFiltered[start:end, fields["isoTime"]], rep(NaN, row-countObs[i]))
-    longitude[,i] <- as.numeric(c(dataBaseFiltered[start:end, fields["lon"]], rep(NaN, row-countObs[i])))
-    latitude[,i] <- as.numeric(c(dataBaseFiltered[start:end, fields["lat"]], rep(NaN, row-countObs[i])))
+    data$names[i] <- dataBaseFiltered[start, fields["names"]]
+    data$seasons[i] <- as.numeric(dataBaseFiltered[start, fields["seasons"]])
+    data$isotimes[,i] <- c(dataBaseFiltered[start:end, fields["isoTime"]], rep(NaN, row-countObs[i]))
+    data$longitude[,i] <- as.numeric(c(dataBaseFiltered[start:end, fields["lon"]], rep(NaN, row-countObs[i])))
+    data$latitude[,i] <- as.numeric(c(dataBaseFiltered[start:end, fields["lat"]], rep(NaN, row-countObs[i])))
     
     if (unitConversion["msw"] == "mph2ms") {
-      msw[,i] <- as.numeric(c(mph2ms(as.numeric(dataBaseFiltered[start:end, fields["msw"]])), rep(NaN, row-countObs[i])))
+      data$msw[,i] <- as.numeric(c(mph2ms(as.numeric(dataBaseFiltered[start:end, fields["msw"]])), rep(NaN, row-countObs[i])))
     } else if (unitConversion["msw"] == "knt2ms") {
-      msw[,i] <- as.numeric(c(knt2ms(as.numeric(dataBaseFiltered[start:end, fields["msw"]])), rep(NaN, row-countObs[i])))
[…]
