### 2023-12-06 e298974
add convert function
