+++ R/defStormsDataset.R
@@ -38,9 +38,72 @@ atm2pa <- function(x) {
 }
 
 
-########
-# Class#
-########
+
+#' Convert variable in the correct metric in the dataset
+#'
+#' @param data database generated with either getDataFromNcdfFile or
+#' getDataFromCsvFile
+#' @param unitConversion named character vector (Cf defStormsDatabase)
+#'
+#' @return data with converted variables
+convertVariables <- function(data, unitConversion){
+  
+  # Maximum sustained wind
+  if (unitConversion["msw"] == "mph2ms") {
+    data$msw <- mph2ms(data$msw)
+  } else if (unitConversion["msw"] == "knt2ms") {
+    data$msw <- knt2ms(data$msw)
+  } else if (unitConversion["msw"] == "kmh2ms") {
+    data$msw <- kmh2ms(data$msw)
+  } 
+  
+  # Radius of maximum wind
+  if ("rmw" %in% names(data)) {
+    if (unitConversion["rmw"] == "nm2km") {
+      data$rmw <- nm2km(data$rmw)
+    } 
+  }
+  
+  # Pressure
+  if ("pressure" %in% names(data)) {
+    if (unitConversion["pressure"] == "mb2pa") {
+      data$pressure <- mb2pa(data$pressure)
+      
+    } else if (unitConversion["pressure"] == "b2pa") {
+      data$pressure <- b2pa(data$pressure)
+      
+    } else if (unitConversion["pressure"] == "psi2pa") {
+      data$pressure <- psi2pa(data$pressure)
+      
+    } else if (unitConversion["pressure"] == "atm2pa") {
+      data$pressure <- atm2pa(data$pressure)
+    } 
+  }
+  
+  # Pressure
+  if ("poci" %in% names(data)) {
+    if (unitConversion["poci"] == "mb2pa") {
+      data$poci <- mb2pa(data$poci)
+      
+    } else if (unitConversion["poci"] == "b2pa") {
+      data$poci <- b2pa(data$poci)
+      
+    } else if (unitConversion["poci"] == "psi2pa") {
+      data$poci <- psi2pa(data$poci)
+      
+    } else if (unitConversion["poci"] == "atm2pa") {
+      data$poci <- atm2pa(data$poci)
+    } 
+  }
+  
+  return(data)
+  
+}
+
+
+#########
+# Class #
+#########
 
 
 #' stormsDataset
@@ -160,11 +223,11 @@ checkInputsdefStormsDataset <- function(filename, fields, basin, seasons, unitCo
   splitedFilename <- strsplit(filename, "\\.")[[1]]
   extension <- splitedFilename[length(splitedFilename)]
   stopifnot("filename must be either a NetCDF (.nc) or a CSV (.csv) file" = extension %in% c("nc", "csv"))
-
+  
   # Checking fields input
   stopifnot("fields must be character" = identical(class(fields), "character"))
   stopifnot("unitConversion must be character" = identical(class(unitConversion), "character"))
-
+  
   # Mandatory fields
   stopifnot("No 'names' selection in fields" = "names" %in% names(fields))
   stopifnot("No 'seasons' selection in fields" = "seasons" %in% names(fields))
@@ -181,7 +244,7 @@ checkInputsdefStormsDataset <- function(filename, fields, basin, seasons, unitCo
   if(extension == "csv"){
     stopifnot("No 'sid' selection in fields" = "sid" %in% names(fields))
   }
-
+  
   # Optional fields
   if (("basin" %in% names(fields)) && is.null(basin)) {
[…]
