+++ man/convertVariables.Rd
@@ -1,0 +1,20 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/defStormsDataset.R
+\name{convertVariables}
+\alias{convertVariables}
+\title{Convert variable in the correct metric in the dataset}
+\usage{
+convertVariables(data, unitConversion)
+}
+\arguments{
+\item{data}{database generated with either getDataFromNcdfFile or
+getDataFromCsvFile}
+
+\item{unitConversion}{named character vector (Cf defStormsDatabase)}
+}
+\value{
+data with converted variables
+}
+\description{
+Convert variable in the correct metric in the dataset
+}
