### 2023-12-06 b77b1c5
Add interactive mode in plotstorms
