+++ R/plotStorms.R
@@ -2,6 +2,42 @@
 
 
 
+#' Get type for icon (leaflet)
+#'
+#' @noRd
+#' @param msw numeric
+#'
+#' @return corresponding icon
+getType <- function(msw) {
+  
+  if (is.na(msw)) {
+    type <- NA
+    
+  } else if (msw < sshs[1]) {
+    type <- "TD"
+    
+  } else if (msw >= sshs[1] & msw < sshs[2]) {
+    type <- "TS"
+    
+  } else if (msw >= sshs[2] & msw < sshs[3]) {
+    type <- "cat1"
+    
+  } else if (msw >= sshs[3] & msw < sshs[4]) {
+    type <- "cat2"
+    
+  } else if (msw >= sshs[4] & msw < sshs[5]) {
+    type <- "cat3"
+    
+  } else if (msw >= sshs[5] & msw < sshs[6]) {
+    type <- "cat4"
+    
+  } else if (msw >= sshs[6]) {
+    type <- "cat5"
+  }
+  
+  return(type)
+}
+
 
 #' Get the right color associated with a wind observation
 #'
@@ -114,9 +150,10 @@ plotLabels <- function(st, by, pos) {
 #' @param loi logical
 #' @param xlim numeric vector
 #' @param ylim numeric vector
+#' @param interactive logical
 #' @return NULL, just stops the function if an error is found
 checkInputsPlotStorms <- function(sts, names, category, labels, by,
-                                  pos, legends, loi, xlim, ylim) {
+                                  pos, legends, loi, xlim, ylim, interactive) {
   # Checking sts input
   stopifnot("no data to plot" = !missing(sts))
 
@@ -169,6 +206,10 @@ checkInputsPlotStorms <- function(sts, names, category, labels, by,
     "legends must be either topright, topleft, bottomleft, bottomright, or none" =
       legends %in% c("topright", "topleft", "bottomleft", "bottomright", "none")
   )
+  
+  #Checking mode input
+  stopifnot("interactive must be logical" = identical(class(interactive), "logical"))
+  stopifnot("interactive must length 1" = length(mode) == 1)
 }
 
 
@@ -207,6 +248,8 @@ checkInputsPlotStorms <- function(sts, names, category, labels, by,
 #' `"bottomleft"`, `"bottomright"`, or `"none"` (legend not plotted).
 #' @param loi logical. Whether (TRUE, default setting) or not (FALSE) to plot the
 #' extent of the buffered location of interest on the map.
+#' @param interactive logical. Whether (FALSE, default setting) or (TRUE) to plot the 
+#' data interactively with leaflet library
 #'
 #' @return A plot of the storm track data.
 #' 
@@ -224,6 +267,10 @@ checkInputsPlotStorms <- function(sts, names, category, labels, by,
 #'
 #' # Plotting Pam over Vanuatu with labels every 6h on the right side of the observations
 #' plotStorms(pam, labels = TRUE, by = 2, pos = 4)
+#' 
+#' # Interactive mode
+#  # plotStorms(pam, interactive=TRUE)
+#' 
 #' }
 #' @export
 plotStorms <- function(sts,
@@ -235,11 +282,10 @@ plotStorms <- function(sts,
                        by = 8,
                        pos = 3,
                        legends = "topleft",
-                       loi = TRUE) {
-  checkInputsPlotStorms(
-    sts, names, category, labels, by, pos, legends,
-    loi, xlim, ylim
-  )
+                       loi = TRUE,
+                       interactive=FALSE) {
+  
+  checkInputsPlotStorms(sts, names, category, labels, by, pos, legends,loi, xlim, ylim, interactive)
 
 
   # Handling spatial extent
@@ -288,54 +334,126 @@ plotStorms <- function(sts,
     }
   }
 
-
[…]
