+++ man/plotStorms.Rd
@@ -14,7 +14,8 @@ plotStorms(
   by = 8,
   pos = 3,
   legends = "topleft",
-  loi = TRUE
+  loi = TRUE,
+  interactive = FALSE
 )
 }
 \arguments{
@@ -55,6 +56,9 @@ when observations are made every 3 (or 6) hours.}
 
 \item{loi}{logical. Whether (TRUE, default setting) or not (FALSE) to plot the
 extent of the buffered location of interest on the map.}
+
+\item{interactive}{logical. Whether (FALSE, default setting) or (TRUE) to plot the
+data interactively with leaflet library}
 }
 \value{
 A plot of the storm track data.
@@ -76,5 +80,8 @@ plotStorms(pam, labels = TRUE)
 
 # Plotting Pam over Vanuatu with labels every 6h on the right side of the observations
 plotStorms(pam, labels = TRUE, by = 2, pos = 4)
+
+# Interactive mode
+
 }
 }
