+++ DESCRIPTION
@@ -26,6 +26,7 @@ Depends:
     R (>= 2.10)
 Imports: 
     graphics,
+    leaflet,
     maps,
     methods,
     ncdf4,
