+++ R/plotBehaviour.R
@@ -16,8 +16,9 @@
 #' @param colorPalette character vector
 #' @param main character
 #' @param legends character
+#' @param interactive logical
 #' @return NULL, just stops the function if an error is found
-checkInputsPlotBehaviour <- function(sts, rasterProduct, xlim, ylim, labels, by, pos, colorPalette, main, legends) {
+checkInputsPlotBehaviour <- function(sts, rasterProduct, xlim, ylim, labels, by, pos, colorPalette, main, legends, interactive) {
   # Checking sts input
   stopifnot("no data to plot" = !missing(sts))
 
@@ -72,6 +73,10 @@ checkInputsPlotBehaviour <- function(sts, rasterProduct, xlim, ylim, labels, by,
     "legends must be either topright, topleft, bottomleft, bottomright, or none" =
       legends %in% c("topright", "topleft", "bottomleft", "bottomright", "none")
   )
+  
+  #Checking mode input
+  stopifnot("interactive must be logical" = identical(class(interactive), "logical"))
+  stopifnot("interactive must length 1" = length(mode) == 1)
 }
 
 
@@ -108,9 +113,10 @@ checkInputsPlotBehaviour <- function(sts, rasterProduct, xlim, ylim, labels, by,
 #' when observations are made every 3 (or 6) hours.
 #' @param pos numeric. If `labels=TRUE`, defines the position of the labels, `1` (above the observation),
 #' `2` (on the left), `3` (below, default setting), and `4` (on the right).
-#' @param legends character. Indicates where to plot the legend, `"topright"`, `"topleft"` (default setting),
+#' @param legends character. Indicates where to plot the legend, `"topright"`(default setting), `"topleft"`,
 #' `"bottomleft"`, `"bottomright"`, or `"none"` (legend not plotted).
-#'
+#' @param interactive logical. Whether (FALSE, default setting) or (TRUE) to plot the 
+#' data interactively with leaflet library
 #' @returns A plot of the storm track data with the raster layer.
 #'
 #' @examples
@@ -139,10 +145,11 @@ plotBehaviour <- function(sts,
                           labels = FALSE,
                           by = 8,
                           pos = 3,
-                          legends = "topleft") {
+                          legends = "topright",
+                          interactive = FALSE) {
   checkInputsPlotBehaviour(
     sts, rasterProduct, xlim, ylim, labels, by, pos, colorPalette,
-    main, legends
+    main, legends, interactive
   )
 
   name <- strsplit(names(rasterProduct), split = "_", fixed = TRUE)[[1]][1]
@@ -170,15 +177,8 @@ plotBehaviour <- function(sts,
     ymin <- ylim[1]
     ymax <- ylim[2]
   }
-
-
-  # Plotting track
-  plotStorms(
-    sts = sts, names = name, xlim = c(xmin, xmax), ylim = c(ymin, ymax),
-    legends = legends
-  )
-
-  # Adding rasterProduct on map
+  
+  # Choose color, range and legends depending on input raster
   if (product == "MSW") {
     col <- mswSSHSPalette
     range <- c(17, 95)
@@ -200,7 +200,7 @@ plotBehaviour <- function(sts,
     range <- c(0, 360)
     leg <- expression(paste("Wind direction (degree)"))
   }
-
+  
   if (!is.null(colorPalette)) {
     col <- colorPalette
   }
@@ -209,74 +209,111 @@ plotBehaviour <- function(sts,
     leg <- main
   }
 
-  # Adding title
-  graphics::title(leg)
-
-
-  terra::plot(rasterProduct,
-    col = col,
-    type = "continuous",
-    xlim = c(xmin - 1, xmax + 1), # we extend W & E by 1°. Needs to be in agreement with plotStorm
[…]
