### 2023-12-07 b342b12
replace interactive with dynamicPlot
