### 2023-12-07 33001a9
fix legend superscript in HTML
