### 2023-12-07 c504724
fix issue with forward pip operators
