### 2023-12-08 f5e0cc2
Minor modifications in verbose mode
