+++ R/spatialBehaviour.R
@@ -1272,12 +1272,7 @@ spatialBehaviour <- function(sts,
     s <- 1 # Initializing count of storms
     cat(" Done\n\n")
     cat("Computation settings:\n")
-    cat("  (*) Temporal resolution: Every", switch(as.numeric(tempRes),
-                                                   "1" = 60,
-                                                   "0.75" = 45,
-                                                   "0.5" = 30,
-                                                   "0.25" = 15
-    ), "min\n")
+    cat("  (*) Temporal resolution: Every", tempRes, "min\n")
     cat("  (*) Space resolution:", names(resolutions[spaceRes]), "\n")
     cat("  (*) Method used:", method, "\n")
     cat("  (*) Product(s) to compute:", product, "\n")
