### 2023-12-08 9f52fd8
Remove unsed function
Co-authored-by: Thomas Arsouze <thomas.arsouze@zoho.com>
