### 2023-12-08 2ba4825
move scale and scalePalette to defStormList
