+++ R/temporalBehaviour.R
@@ -4,7 +4,6 @@
 #'
 #' @noRd
 #' @param sts StormsList object
-#' @param sds `stormsDataset` object
 #' @param points data.frame
 #' @param product character
 #' @param windThreshold numeric
@@ -14,14 +13,11 @@
 #' @param tempRes numeric
 #' @param verbose logical
 #' @return NULL
-checkInputsTemporalBehaviour <- function(sts, sds, points, product, windThreshold, method, asymmetry,
+checkInputsTemporalBehaviour <- function(sts, points, product, windThreshold, method, asymmetry,
                                          empiricalRMW, tempRes, verbose) {
   # Checking sts input
   stopifnot("no data found" = !missing(sts))
   
-  # Checking sts input
-  stopifnot("no data found" = !missing(sds))
-
   # Checking points input
   stopifnot("no data found" = !missing(points))
   stopifnot("points must be data.frame" = identical(class(points), "data.frame"))
@@ -212,7 +208,6 @@ finalizeResult <- function(finalResult, result, product, points, isoT, indices, 
 #' It also allows to compute three associated summary statistics.
 #'
 #' @param sts `StormsList` object.
-#' @param sds `stormsDataset` object used to generate `sts` input
 #' @param points data.frame. Consisting of two columns names as "x" (for the longitude) and
 #' "y" (for the latitude), providing the coordinates in decimal degrees of the point locations. Row names
 #' can also be provided to named the locations.
@@ -410,19 +405,18 @@ finalizeResult <- function(finalResult, result, product, points, isoT, indices, 
 #'
 #' # Computing time series of wind speed and direction for Pam
 #' # over points 1 and 2 defined above
-#' ts.pam <- temporalBehaviour(pam, sds, points = pts)
+#' ts.pam <- temporalBehaviour(pam, points = pts)
 #'
 #' # Computing PDI for Pam over points 1 and 2 defined above
-#' pdi.pam <- temporalBehaviour(pam, sds, points = pts, product = "PDI")
+#' pdi.pam <- temporalBehaviour(pam, points = pts, product = "PDI")
 #'
 #' # Computing the duration of exposure to wind speeds above the thresholds
 #' # used by the Saffir-Simpson hurricane wind scale for Pam
 #' # over points 1 and 2 defined above
-#' exp.pam <- temporalBehaviour(pam, sds, points = pts, product = "Exposure")
+#' exp.pam <- temporalBehaviour(pam, points = pts, product = "Exposure")
 #' }
 #' @export
 temporalBehaviour <- function(sts,
-                              sds,
                               points,
                               product = "TS",
                               windThreshold = c(18, 33, 42, 49, 58, 70),
@@ -433,10 +427,10 @@ temporalBehaviour <- function(sts,
                               verbose = 1) {
   
   if(is.null(windThreshold)){
-    windThreshold = sshs
+    windThreshold = sts@scale
   }
   
-  checkInputsTemporalBehaviour(sts, sds, points, product, windThreshold, method, asymmetry, empiricalRMW, tempRes, verbose)
+  checkInputsTemporalBehaviour(sts, points, product, windThreshold, method, asymmetry, empiricalRMW, tempRes, verbose)
 
 
   if (verbose > 0) {
