+++ R/plotStorms.R
@@ -138,7 +138,7 @@ checkInputsPlotStorms <- function(sts, names, category, labels, by,
     stopifnot("xlim must be numeric" = identical(class(xlim), "numeric"))
     stopifnot("xlim must length 2" = length(xlim) == 2)
     stopifnot("xlim must have valid longitude coordinates" = xlim >= 0 &
-      xlim <= 360)
+                xlim <= 360)
   }
 
   # Checking ylim input
@@ -146,7 +146,7 @@ checkInputsPlotStorms <- function(sts, names, category, labels, by,
     stopifnot("ylim must be numeric" = identical(class(ylim), "numeric"))
     stopifnot("ylim must length 2" = length(ylim) == 2)
     stopifnot("ylim must have valid latitude coordinates" = ylim >= -90 &
-      ylim <= 90)
+                ylim <= 90)
   }
 
   # Checking logical inputs
@@ -170,7 +170,7 @@ checkInputsPlotStorms <- function(sts, names, category, labels, by,
     "legends must be either topright, topleft, bottomleft, bottomright, or none" =
       legends %in% c("topright", "topleft", "bottomleft", "bottomright", "none")
   )
-  
+
   #Checking mode input
   stopifnot("dynamicPlot must be logical" = identical(class(dynamicPlot), "logical"))
   stopifnot("dynamicPlot must length 1" = length(mode) == 1)
@@ -231,10 +231,10 @@ checkInputsPlotStorms <- function(sts, names, category, labels, by,
 #'
 #' # Plotting Pam over Vanuatu with labels every 6h on the right side of the observations
 #' plotStorms(pam, labels = TRUE, by = 2, pos = 4)
-#' 
+#'
 #' # dynamicPlot mode
 #  # plotStorms(pam, dynamicPlot=TRUE)
-#' 
+#'
 #' }
 #' @export
 plotStorms <- function(sts,
@@ -247,8 +247,8 @@ plotStorms <- function(sts,
                        pos = 3,
                        legends = "topright",
                        loi = TRUE,
-                       dynamicPlot=FALSE) {
-  
+                       dynamicPlot = FALSE) {
+
   checkInputsPlotStorms(sts, names, category, labels, by, pos, legends,loi, xlim, ylim, dynamicPlot)
 
 
@@ -298,104 +298,109 @@ plotStorms <- function(sts,
     }
   }
 
-  if(!dynamicPlot){
-    
+  if (!dynamicPlot) {
+
     # Plotting base map
     world <- rworldmap::getMap(resolution = "high")
-    
+
     maps::map(world,
-              fill = TRUE,
-              col = groundColor,
-              bg = oceanColor,
-              wrap = c(0, 360),
-              xlim = c(ext$xmin - 1, ext$xmax + 1), # we extend W & E by 1°
-              ylim = c(ext$ymin - 1, ext$ymax + 1)
+      fill = TRUE,
+      col = groundColor,
+      bg = oceanColor,
+      wrap = c(0, 360),
+      xlim = c(ext$xmin - 1, ext$xmax + 1), # we extend W & E by 1°
+      ylim = c(ext$ymin - 1, ext$ymax + 1)
     ) # we extend S & N by 1°
     maps::map.axes(cex.axis = 1)
-    
-    
+
+
     # Plotting loi
     if (loi) {
       plot(sts@spatialLoiBuffer, lwd = 1, border = "blue", add = TRUE)
     }
-    
+
     if (loi && as.character(sf::st_geometry_type(sts@spatialLoi)) == "POINT") {
[…]
