### 2023-12-08 8b3c275
default windthreshold in temporalbehaviour
