+++ R/temporalBehaviour.R
@@ -219,7 +219,7 @@ finalizeResult <- function(finalResult, result, product, points, isoT, indices, 
 #'   }
 #' @param windThreshold numeric vector. Minimal wind threshold(s) (in \eqn{m.s^{-1}}) used to
 #'   compute the duration of exposure when `product="Exposure"`. Default value is to set NULL, in this 
-#'   case, the windthresholds are the one used in the scale defined in the stormdatabase 
+#'   case, the windthresholds are the one used in the scale defined in the stormdatabase.
 #' @param method character. Model used to compute wind speed and direction.
 #' Three different models are implemented:
 #'   \itemize{
@@ -419,7 +419,7 @@ finalizeResult <- function(finalResult, result, product, points, isoT, indices, 
 temporalBehaviour <- function(sts,
                               points,
                               product = "TS",
-                              windThreshold = c(18, 33, 42, 49, 58, 70),
+                              windThreshold = NULL,
                               method = "Willoughby",
                               asymmetry = "Chen",
                               empiricalRMW = FALSE,
