### 2023-12-08 e19e41d
sshsPalette
Co-authored-by: Thomas Arsouze <thomas.arsouze@zoho.com>
