+++ R/defStormsList.R
@@ -1042,7 +1042,7 @@ defStormsList <- function(sds,
                    names = NULL,
                    maxDist = 300,
                    scale = sshs,
-                   scalePalette = SSHS_PALETTE,
+                   scalePalette = sshsPalette,
                    removeUnder = NULL,
                    verbose = 2) {
 
