### 2023-12-08 379d6c4
remove SSHS_PALETTE from internalData
