### 2023-12-08 2231d65
remove SSHS_PALETTE
