### 2023-12-08 f886b36
remove SSHS_PALETTE
Co-authored-by: Thomas Arsouze <thomas.arsouze@zoho.com>
