+++ man/spatialBehaviour.Rd
@@ -29,7 +29,7 @@ spatialBehaviour(
 
 \item{windThreshold}{numeric vector. Minimal wind threshold(s) (in \eqn{m.s^{-1}}) used to
 compute the duration of exposure when \code{product="Exposure"}. Default value is to set NULL, in this
-case, the windthresholds are the one used in the scale defined in the stormdatabase}
+case, the windthresholds are the one used in the scale defined in the stormdatabase.}
 
 \item{method}{character. Model used to compute wind speed and direction.
 Three different models are implemented:
