### 2023-12-09 01e5424
default parameters for scalePalette
