+++ man/defStormsList.Rd
@@ -11,7 +11,7 @@ defStormsList(
   names = NULL,
   maxDist = 300,
   scale = sshs,
-  scalePalette = sshsPalette,
+  scalePalette = NULL,
   removeUnder = NULL,
   verbose = 2
 )
