+++ R/temporalBehaviour.R
@@ -393,7 +393,7 @@ finalizeResult <- function(finalResult, result, product, points, isoT, indices, 
 #' Monthly Weather Review, 134(4), 1102–1120. https://doi.org/10.1175/MWR3106.1
 #'
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' # Creating a stormsDataset
 #' sds <- defStormsDataset()
 #'
