### 2023-12-12 9c18c08
Default scalePalette values
