### 2023-12-12 68645fe
handle notNamed storms in database
