### 2023-12-13 bb1378d
Remove NA lon-lat lines in csv routine
