+++ tests/testthat/test-stormsDataset.R
@@ -3,10 +3,12 @@
 
 
 
+
 test_that("Test checkInputsdefStormsDataset function", {
   # Checking filename input
   expect_error(
     checkInputsdefStormsDataset(
+      sep = NULL,
       fields = c(
         "basin" = "basin",
         "names" = "name",
@@ -22,6 +24,7 @@ test_that("Test checkInputsdefStormsDataset function", {
       ),
       basin = "SP",
       seasons = c(1980, as.numeric(format(Sys.time(
+        
       ), "%Y"))),
       unitConversion = c(
         msw = "knt2ms",
@@ -37,6 +40,7 @@ test_that("Test checkInputsdefStormsDataset function", {
   expect_error(
     checkInputsdefStormsDataset(
       filename = 1,
+      sep = NULL,
       fields = c(
         "basin" = "basin",
         "names" = "name",
@@ -52,6 +56,7 @@ test_that("Test checkInputsdefStormsDataset function", {
       ),
       basin = "SP",
       seasons = c(1980, as.numeric(format(Sys.time(
+        
       ), "%Y"))),
       unitConversion = c(
         msw = "knt2ms",
@@ -67,6 +72,7 @@ test_that("Test checkInputsdefStormsDataset function", {
   expect_error(
     checkInputsdefStormsDataset(
       filename = TRUE,
+      sep = NULL,
       fields = c(
         "basin" = "basin",
         "names" = "name",
@@ -82,6 +88,7 @@ test_that("Test checkInputsdefStormsDataset function", {
       ),
       basin = "SP",
       seasons = c(1980, as.numeric(format(Sys.time(
+        
       ), "%Y"))),
       unitConversion = c(
         msw = "knt2ms",
@@ -98,6 +105,7 @@ test_that("Test checkInputsdefStormsDataset function", {
   expect_error(
     checkInputsdefStormsDataset(
       filename = c("data", ".nc"),
+      sep = NULL,
       fields = c(
         "basin" = "basin",
         "names" = "name",
@@ -113,6 +121,7 @@ test_that("Test checkInputsdefStormsDataset function", {
       ),
       basin = "SP",
       seasons = c(1980, as.numeric(format(Sys.time(
+        
       ), "%Y"))),
       unitConversion = c(
         msw = "knt2ms",
@@ -124,14 +133,116 @@ test_that("Test checkInputsdefStormsDataset function", {
       verbose = 1
     )
   )
+  
+  # Checking sep input
+  expect_error(
+    checkInputsdefStormsDataset(
+      filename = "database.nc",
+      sep = 1,
+      fields = c(
+        "basin" = "basin",
+        "names" = "name",
+        "seasons" = "season",
+        "isoTime" = "iso_time",
+        "lon" = "usa_lon",
+        "lat" = "usa_lat",
+        "msw" = "usa_wind",
+        "sshs" = "usa_sshs",
+        "rmw" = "usa_rmw",
+        "pressure" = "usa_pres",
+        "poci" = "usa_poci"
+      ),
+      basin = "SP",
+      seasons = c(1980, as.numeric(format(Sys.time(
+        
+      ), "%Y"))),
+      unitConversion = c(
+        msw = "knt2ms",
+        rmw = "nm2km",
+        pressure = "mb2pa",
+        poci = "mb2pa"
+      ),
+      notNamed = "NOT_NAMED",
+      verbose = 1
+    )
+  )
+  
+  expect_error(
+    checkInputsdefStormsDataset(
+      filename = "database.nc",
+      sep = TRUE,
+      fields = c(
+        "basin" = "basin",
+        "names" = "name",
+        "seasons" = "season",
+        "isoTime" = "iso_time",
+        "lon" = "usa_lon",
+        "lat" = "usa_lat",
[…]
