+++ R/defStormsDataset.R
@@ -677,8 +677,14 @@ getDataFromCsvFile <- function(filename, sep, fields, basin, seasons, unitConver
 #' @examples
 #' # Creating a `stormsDataset` object with storms between 2010 and 2015
 #' # in the South Pacific using the NetCDF provided with the package
-#' SP_2015_2020 <- defStormsDataset(seasons = c(2010, 2015))
-#' str(SP_2015_2020)
+#' SP_2015_2020_nc <- defStormsDataset(seasons = c(2010, 2015))
+#' str(SP_2015_2020_nc)
+#' 
+#' # Creating a `stormsDataset` object with storms between 2010 and 2015
+#' # in the South Pacific using the CSV provided with the package
+#' fileName <- system.file("extdata", "test_dataset.csv", package = "StormR")
+#' SP_2015_2020_csv <- defStormsDataset(seasons = c(2010, 2021))
+#' str(SP_2015_2020_csv)
 #' @export
 defStormsDataset <- function(filename = system.file("extdata", "test_dataset.nc", package = "StormR"),
                              sep = NULL,
