+++ tests/testthat/test-stormsList.R
@@ -1,293 +1,472 @@
 
 
 
+
 test_that("Tests invalid inputs", {
   suppressWarnings(sds <- defStormsDataset(verbose = 0))
-
+  
   # Checking sdb input
-  expect_error(checkInputsDefStormsList(
-    loi = NULL,
-    seasons = c(1980, 2022),
-    names = "PAM",
-    maxDist = 300,
-    verbose = 2,
-    removeUnder = NULL
-  ))
-
+  expect_error(
+    checkInputsDefStormsList(
+      loi = NULL,
+      seasons = c(1980, 2022),
+      names = "PAM",
+      maxDist = 300,
+      scale = sshs,
+      scalePalette = NULL,
+      scale = sshs,
+      scalePalette = NULL,
+      verbose = 2,
+      removeUnder = NULL
+    )
+  )
+  
   # Checking loi input
-  expect_error(checkInputsDefStormsList(
-    sds = sds,
-    seasons = c(1980, 2022),
-    names = "PAM",
-    maxDist = 300,
-    verbose = 2,
-    removeUnder = NULL
-  ))
-
-  expect_error(checkInputsDefStormsList(
-    sds = sds,
-    loi = c("Vanuatu", "New Caledonia"),
-    seasons = c(1980, 2022),
-    names = "PAM",
-    maxDist = 300,
-    verbose = 2,
-    removeUnder = NULL
-  ))
-
-  expect_error(checkInputsDefStormsList(
-    sds = sds,
-    loi = 169,
-    seasons = c(1980, 2022),
-    names = "PAM",
-    maxDist = 300,
-    verbose = 2,
-    removeUnder = NULL
-  ))
-
-  expect_error(checkInputsDefStormsList(
-    sds = sds,
-    loi = c(169, 170, -15, -14),
-    seasons = c(1980, 2022),
-    names = "PAM",
-    maxDist = 300,
-    verbose = 2,
-    removeUnder = NULL
-  ))
-
-  expect_error(checkInputsDefStormsList(
-    sds = sds,
-    loi = c(169, 170, -15, -14),
-    seasons = c(1980, 2022),
-    names = "PAM",
-    maxDist = 300,
-    verbose = 2,
-    removeUnder = NULL
-  ))
-
-  expect_error(checkInputsDefStormsList(
-    sds = sds,
-    loi = c(169, -100),
-    seasons = c(1980, 2022),
-    names = "PAM",
-    maxDist = 300,
-    verbose = 2,
-    removeUnder = NULL
-  ))
-
-
+  expect_error(
+    checkInputsDefStormsList(
+      sds = sds,
+      seasons = c(1980, 2022),
+      names = "PAM",
+      maxDist = 300,
+      scale = sshs,
+      scalePalette = NULL,
+      verbose = 2,
+      removeUnder = NULL
+    )
+  )
+  
+  expect_error(
+    checkInputsDefStormsList(
+      sds = sds,
+      loi = c("Vanuatu", "New Caledonia"),
+      seasons = c(1980, 2022),
+      names = "PAM",
+      maxDist = 300,
+      scale = sshs,
+      scalePalette = NULL,
+      verbose = 2,
+      removeUnder = NULL
+    )
+  )
+  
+  expect_error(
+    checkInputsDefStormsList(
+      sds = sds,
+      loi = 169,
+      seasons = c(1980, 2022),
+      names = "PAM",
+      maxDist = 300,
+      scale = sshs,
+      scalePalette = NULL,
+      verbose = 2,
+      removeUnder = NULL
+    )
+  )
+  
+  expect_error(
+    checkInputsDefStormsList(
+      sds = sds,
+      loi = c(169, 170,-15,-14),
[…]
