### 2023-12-15 #5766 @gsapijaszko
Yes, I have installed and tested the software. Otherwise I wouldn't mark check marks in review. Yes, I went through examples/help and played with the data. I don't have anything to complain about, neither paper, neither software.

Regards,
Grzegorz
