### 2023-12-15 fcebb98
few bug fix + documentation
