+++ R/plotStorms.R
@@ -131,7 +131,7 @@ checkInputsPlotStorms <- function(sts, names, category, labels, by,
   # Checking category input
   if (!is.null(category)) {
     stopifnot("category must be numeric(s)" = identical(class(category), "numeric"))
-    stopifnot("Invalid category input" = category %in% c(-1, -2, 0, 1, 2, 3, 4, 5))
+    stopifnot("Invalid category input" = category %in% seq(0, length(sts@scale)))
   }
 
   # Checking xlim input
@@ -184,18 +184,9 @@ checkInputsPlotStorms <- function(sts, names, category, labels, by,
 #' @param sts `StormsList` object
 #' @param names character vector. Name(s) of the storm(s) in capital letters.
 #'  If `names = NULL` (default setting), all storms are plotted.
-#' @param category numeric vector. Category of storms to be plotted in the Saffir-Simpson hurricane wind scale.
-#'  Can be a single category or a range of categories among:
-#' \itemize{
-#'    \item -1, for tropical depression,
-#'    \item 0, for tropical storms,
-#'    \item 1, for category 1 tropical cyclone,
-#'    \item 2, for category 2 tropical cyclone,
-#'    \item 3, for category 3 tropical cyclone,
-#'    \item 4, for category 4 tropical cyclone, or
-#'    \item 5, for category 5 tropical cyclone.
-#' }
-#'  If `category=NULL` (default setting), all storms are plotted.
+#' @param category numeric vector. Category of storms to be plotted among the level in
+#'  the windscale provided in `sts` input. If `category=NULL` (default setting),
+#'  all storms are plotted.
 #' @param xlim numeric vector. The x limits of the plot.
 #' @param ylim numeric vector. The y limits of the plot.
 #' @param labels logical. Whether (TRUE) or not (FALSE, default setting) to add labels with the name
