+++ pkgdown/_pkgdown.yml
@@ -65,6 +65,7 @@ reference:
   contents:
   - '`plotStorms`'
   - '`plotBehaviour`'
+  - '`plotTemporal`'
 
 
 - title: Write
