+++ tests/testthat/test-stormsList.R
@@ -9,7 +9,7 @@ test_that("Tests invalid inputs", {
   expect_error(
     checkInputsDefStormsList(
       loi = NULL,
-      seasons = c(1980, 2022),
+      seasons = c(2015, 2021),
       names = "PAM",
       maxDist = 300,
       scale = sshs,
@@ -25,7 +25,7 @@ test_that("Tests invalid inputs", {
   expect_error(
     checkInputsDefStormsList(
       sds = sds,
-      seasons = c(1980, 2022),
+      seasons = c(2015, 2021),
       names = "PAM",
       maxDist = 300,
       scale = sshs,
@@ -39,7 +39,7 @@ test_that("Tests invalid inputs", {
     checkInputsDefStormsList(
       sds = sds,
       loi = c("Vanuatu", "New Caledonia"),
-      seasons = c(1980, 2022),
+      seasons = c(2015, 2021),
       names = "PAM",
       maxDist = 300,
       scale = sshs,
@@ -53,7 +53,7 @@ test_that("Tests invalid inputs", {
     checkInputsDefStormsList(
       sds = sds,
       loi = 169,
-      seasons = c(1980, 2022),
+      seasons = c(2015, 2021),
       names = "PAM",
       maxDist = 300,
       scale = sshs,
@@ -67,7 +67,7 @@ test_that("Tests invalid inputs", {
     checkInputsDefStormsList(
       sds = sds,
       loi = c(169, 170,-15,-14),
-      seasons = c(1980, 2022),
+      seasons = c(2015, 2021),
       names = "PAM",
       maxDist = 300,
       scale = sshs,
@@ -81,7 +81,7 @@ test_that("Tests invalid inputs", {
     checkInputsDefStormsList(
       sds = sds,
       loi = c(169, 170,-15,-14),
-      seasons = c(1980, 2022),
+      seasons = c(2015, 2021),
       names = "PAM",
       maxDist = 300,
       scale = sshs,
@@ -95,7 +95,7 @@ test_that("Tests invalid inputs", {
     checkInputsDefStormsList(
       sds = sds,
       loi = c(169,-100),
-      seasons = c(1980, 2022),
+      seasons = c(2015, 2021),
       names = "PAM",
       maxDist = 300,
       scale = sshs,
@@ -226,7 +226,7 @@ test_that("Tests invalid inputs", {
     checkInputsDefStormsList(
       sds = sds,
       loi = "Vanuatu",
-      seasons = c(1980, 2020),
+      seasons = c(2015, 2021),
       names = "PAM",
       maxDist = "hui",
       verbose = 2,
@@ -238,7 +238,7 @@ test_that("Tests invalid inputs", {
     checkInputsDefStormsList(
       sds = sds,
       loi = "Vanuatu",
-      seasons = c(1980, 2020),
+      seasons = c(2015, 2021),
       names = "PAM",
       maxDist = c(1, 3),
       verbose = 2,
@@ -250,7 +250,7 @@ test_that("Tests invalid inputs", {
     checkInputsDefStormsList(
       sds = sds,
       loi = "Vanuatu",
-      seasons = c(1980, 2020),
+      seasons = c(2015, 2021),
       names = "PAM",
       maxDist = -10,
       verbose = 2,
@@ -262,7 +262,7 @@ test_that("Tests invalid inputs", {
     checkInputsDefStormsList(
       sds = sds,
       loi = "Vanuatu",
-      seasons = c(1980, 2020),
+      seasons = c(2015, 2021),
       names = "PAM",
       maxDist = TRUE,
       verbose = 2,
@@ -275,7 +275,7 @@ test_that("Tests invalid inputs", {
     checkInputsDefStormsList(
[…]
