### 2023-12-17 3a7933d
bug fix + tests
