+++ R/plotBehaviour.R
@@ -76,7 +76,7 @@ checkInputsPlotBehaviour <- function(sts, rasterProduct, xlim, ylim, labels, by,
   
   #Checking mode input
   stopifnot("dynamicPlot must be logical" = identical(class(dynamicPlot), "logical"))
-  stopifnot("dynamicPlot must length 1" = length(mode) == 1)
+  stopifnot("dynamicPlot must length 1" = length(dynamicPlot) == 1)
 }
 
 
