+++ tests/testthat/test-plotBehaviour.R
@@ -2,392 +2,549 @@
 
 
 
+
 test_that("Tests checkInputsPlotBehaviour function", {
   suppressWarnings(sds <- defStormsDataset(verbose = 0))
-  pam <- defStormsList(sds, loi = "Vanuatu", names = "PAM", verbose = 0)
+  pam <-
+    defStormsList(sds,
+                  loi = "Vanuatu",
+                  names = "PAM",
+                  verbose = 0)
   msw <- spatialBehaviour(pam, verbose = 0)
-
+  
   # Checking sts input
-  expect_error(checkInputsPlotBehaviour(
-    raster_product = msw,
-    color_palette = NULL,
-    main = NULL,
-    legends = "topleft",
-    xlim = NULL,
-    ylim = NULL,
-    labels = FALSE,
-    by = 8,
-    pos = 3
-  ))
-
+  expect_error(
+    checkInputsPlotBehaviour(
+      raster_product = msw,
+      color_palette = NULL,
+      main = NULL,
+      legends = "topleft",
+      xlim = NULL,
+      ylim = NULL,
+      labels = FALSE,
+      by = 8,
+      pos = 3,
+      dynamicPLot = FALSE
+    )
+  )
+  
   # Checking raster_product input
-  expect_error(checkInputsPlotBehaviour(
-    sts = pam,
-    color_palette = NULL,
-    main = NULL,
-    legends = "topleft",
-    xlim = NULL,
-    ylim = NULL,
-    labels = FALSE,
-    by = 8,
-    pos = 3
-  ))
-
+  expect_error(
+    checkInputsPlotBehaviour(
+      sts = pam,
+      color_palette = NULL,
+      main = NULL,
+      legends = "topleft",
+      xlim = NULL,
+      ylim = NULL,
+      labels = FALSE,
+      by = 8,
+      pos = 3,
+      dynamicPLot = FALSE
+    )
+  )
+  
   # Checking color_palette input
-  expect_error(checkInputsPlotBehaviour(
-    sts = pam,
-    raster_product = msw,
-    color_palette = TRUE,
-    main = NULL,
-    legends = "topleft",
-    xlim = NULL,
-    ylim = NULL,
-    labels = FALSE,
-    by = 8,
-    pos = 3
-  ))
-
-  expect_error(checkInputsPlotBehaviour(
-    sts = pam,
-    raster_product = msw,
-    color_palette = c(1, 2, 3),
-    main = NULL,
-    legends = "topleft",
-    xlim = NULL,
-    ylim = NULL,
-    labels = FALSE,
-    by = 8,
-    pos = 3
-  ))
-
+  expect_error(
+    checkInputsPlotBehaviour(
+      sts = pam,
+      raster_product = msw,
+      color_palette = TRUE,
+      main = NULL,
+      legends = "topleft",
+      xlim = NULL,
+      ylim = NULL,
+      labels = FALSE,
+      by = 8,
+      pos = 3,
+      dynamicPLot = FALSE
+    )
+  )
+  
+  expect_error(
+    checkInputsPlotBehaviour(
+      sts = pam,
+      raster_product = msw,
+      color_palette = c(1, 2, 3),
+      main = NULL,
+      legends = "topleft",
+      xlim = NULL,
+      ylim = NULL,
+      labels = FALSE,
+      by = 8,
+      pos = 3,
+      dynamicPLot = FALSE
+    )
+  )
+  
   # Checking xlim/ylim input
-  expect_error(checkInputsPlotBehaviour(
-    sts = pam,
-    raster_product = msw,
-    color_palette = NULL,
-    main = NULL,
-    legends = "topleft",
-    xlim = "132",
-    ylim = NULL,
[…]
