+++ tests/testthat/test-plotStorms.R
@@ -3,22 +3,15 @@
 
 
 test_that("Test getColors function", {
-  # TODO ouput is now named --> how to proceed ?
-  expect_error(getColors(50))
-  expect_equal(getColors(NA, sshs, SSHS_PALETTE), NA)
-  # expect_equal(getColors(10, sshs, SSHS_PALETTE), "#00CCFF")
-  # expect_equal(getColors(18, sshs, SSHS_PALETTE), "#00CCCC")
-  # expect_equal(getColors(25, sshs, SSHS_PALETTE), "#00CCCC")
-  # expect_equal(getColors(33, sshs, SSHS_PALETTE), "#FFFFB2")
-  # expect_equal(getColors(36, sshs, SSHS_PALETTE), "#FFFFB2")
-  # expect_equal(getColors(42, sshs, SSHS_PALETTE), "#FECC5C")
-  # expect_equal(getColors(46, sshs, SSHS_PALETTE), "#FECC5C")
-  # expect_equal(getColors(49, sshs, SSHS_PALETTE), "#FD8D3C")
-  # expect_equal(getColors(53, sshs, SSHS_PALETTE), "#FD8D3C")
-  # expect_equal(getColors(58, sshs, SSHS_PALETTE), "#F03B20")
-  # expect_equal(getColors(65, sshs, SSHS_PALETTE), "#F03B20")
-  # expect_equal(getColors(70, sshs, SSHS_PALETTE), "#BD0026")
-  # expect_equal(getColors(100, sshs, SSHS_PALETTE), "#BD0026")
+  
+  
+  scale <- c(10,30,50)
+  scalePalette <- c("red", "green", "blue", "black")
+  expect_equal(getColors(5, scale, scalePalette), "red")
+  expect_equal(getColors(25, scale, scalePalette), "green")
+  expect_equal(getColors(45, scale, scalePalette), "blue")
+  expect_equal(getColors(65, scale, scalePalette), "black")
+  
 })
 
 
