### 2023-12-19 2a2ae7d
fix getColors
