### 2023-12-19 49baa74
fix getColors
