+++ src/spatialBehaviour.cpp
@@ -1,0 +1,4 @@
+#include <Rcpp.h>
+using namespace Rcpp;
+
+
