+++ DESCRIPTION
@@ -29,6 +29,7 @@ Imports:
     maps,
     methods,
     ncdf4,
+    Rcpp,
     rworldmap,
     sf,
     stringr,
@@ -47,3 +48,5 @@ Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.2.3
+LinkingTo: 
+    Rcpp
