### 2023-12-19 d02dcfa
add rcpp functions
