+++ src/spatialBehaviour.cpp
@@ -1,4 +1,263 @@
 #include <Rcpp.h>
+#include <cmath>
+
 using namespace Rcpp;
+using namespace std;
+
+
+
+// =============================================================================
+// =====================ccp functions for models ===============================
+// =============================================================================
+
+// [[Rcpp::export]]
+NumericVector
+  willoughby_cpp(NumericVector r,
+                 double rmw,
+                 double msw,
+                 double lat) {
+    int s = r.size();
+    NumericVector vr(s);
+    double x2 = 25;
+    double x1 = 287.6 - 1.942 * msw + 7.799 * log(rmw) + 1.819 * abs(lat);
+    double a = 0.5913 + 0.0029 * msw - 0.1361 * log(rmw) - 0.0042 * abs(lat);
+    double n = 2.1340 + 0.0077 * msw - 0.4522 * log(rmw) - 0.0038 * abs(lat);
+    for(int i = 0; i < s; ++i) {
+      if (r[i] >= rmw) {
+        vr[i] = msw * ((1 - a) * exp(-abs((r[i] - rmw) / x1)) +
+          a * exp(-abs(r[i] - rmw) / x2));
+      } else {
+        vr[i] = msw * abs(pow(r[i] / rmw,n));
+      }
+      vr[i] = std::round(vr[i]* 1000.0) / 1000.0;
+    }
+    return vr;
+  }
+
+
+// [[Rcpp::export]]
+NumericVector
+  holland_cpp(NumericVector r,
+              double rmw,
+              double msw,
+              double pc,
+              double poci,
+              double lat){ //warning here
+
+    double rho = 1.15; //air density
+    double f = 2 * 7.29 * pow(10,-5) * sin(lat); //Coriolis parameter
+    double b = rho * exp(1) * pow(msw, 2) / (poci - pc);
+
+    int s = r.size();
+    NumericVector vr(s);
+
+    for(int i = 0; i < s; ++i) {
+      vr[i] = sqrt(b / rho * pow(rmw / r[i], b) * (poci - pc) * exp(-pow(rmw / r[i], b)) + pow(r[i] * f / 2, 2)) - r[i] * f / 2;
+      vr[i] = std::round(vr[i]* 1000.0) / 1000.0;
+    }
+
+    return vr;
+
+  }
+
+
+// [[Rcpp::export]]
+NumericVector
+  boose_cpp(NumericVector r,
+            double rmw,
+            double msw,
+            double pc,
+            double poci,
+            double x,
+            double y,
+            double vx,
+            double vy,
+            double vh,
+            double landIntersect,
+            double lat){
+
+
+    double rho = 1; //air density
+    double b = rho * exp(1) * pow(msw, 2) / (poci - pc);
+
+    int s = r.size();
+    NumericVector vr(s);
+    NumericVector angle(s);
+
+    if(lat>= 0){
+      // Northern Hemisphere, t is clockwise
+      angle = atan2(vy, vx) - atan2(y, x);
+    }else{
+      // Southern Hemisphere, t is counterclockwise
+      angle = atan2(y, x) - atan2(vy, vx);
+    }
+
+    for(int i = 0; i < s; ++i) {
+      double v = sqrt(pow(rmw / r[i], b) * exp(1 - pow(rmw / r[i], b)));
+
+      if(landIntersect == 1){
+        vr[i] = 0.8 * (msw - (1 - sin(angle[i])) * vh / 2) * v;
+      }else{
+        vr[i] = (msw - (1 - sin(angle[i])) * vh / 2) * v;
+      }
+
+      vr[i] = std::round(vr[i]* 1000.0) / 1000.0;
+    }
+
+    return vr;
+  }
+
+// =============================================================================
[…]
