+++ R/RcppExports.R
@@ -1,0 +1,15 @@
+# Generated by using Rcpp::compileAttributes() -> do not edit by hand
+# Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
+
+willoughby_cpp <- function(r, rmw, msw, lat) {
+    .Call('_StormR_willoughby_cpp', PACKAGE = 'StormR', r, rmw, msw, lat)
+}
+
+holland_cpp <- function(r, rmw, msw, pc, poci, lat) {
+    .Call('_StormR_holland_cpp', PACKAGE = 'StormR', r, rmw, msw, pc, poci, lat)
+}
+
+boose_cpp <- function(r, rmw, msw, pc, poci, x, y, vx, vy, vh, landIntersect, lat) {
+    .Call('_StormR_boose_cpp', PACKAGE = 'StormR', r, rmw, msw, pc, poci, x, y, vx, vy, vh, landIntersect, lat)
+}
+
