### 2023-12-19 7b9b437
add librairies for parallel computing
