+++ R/spatialBehaviour.R
@@ -986,6 +986,164 @@ maskProduct <- function(finalStack, loi, template) {
 }
 
 
+###########################
+# main processing routine #
+###########################
+
+processingStorm <- function(storm,
+                            product,
+                            method,
+                            asymmetry,
+                            empiricalRMW,
+                            tempRes,
+                            spaceRes,
+                            windThresholds,
+                            rasterTemplate,
+                            stackMSW,
+                            stackPDI,
+                            stackEXP,
+                            stackWIND,
+                            count,
+                            totalStorms,
+                            verbose){
+  
+  # Buffer size in degree
+  buffer <- 2.5
+  
+  # Handling indices inside loi.buffer or not
+  ind <- getIndices(storm, 2, product)
+  
+  it1 <- storm@obs.all$iso.time[1]
+  it2 <- storm@obs.all$iso.time[2]
+  timeDiff <- as.numeric(as.POSIXct(it2) - as.POSIXct(it1))
+  # Interpolated time step dt, default value dt <- 4 --> 1h
+  dt <- 1 + (1 / tempRes * timeDiff) # + 1 for the limit values
+  
+  # Getting data associated with storm st
+  dataTC <- getDataInterpolate(storm, ind, dt, timeDiff, empiricalRMW, method)
+  
+  nbStep <- dim(dataTC)[1] - 1
+  
+  if (verbose > 0) {
+    step <- 1
+    cat(storm@name, " (", count, "/", totalStorms, ")\n")
+    pb <- utils::txtProgressBar(min = step, max = nbStep, style = 3)
+  }
+  
+  auxStackMSW <- c()
+  auxStackPDI <- c()
+  auxStackEXP <- c()
+  auxStackSpeed <- c()
+  auxStackDirection <- c()
+  
+  for (j in 1:nbStep) {
+    # Making template to compute wind profiles
+    rasterTemplateModel <- makeTemplateModel(rasterTemplate, buffer, dataTC, j)
+    rasterWind <- rasterTemplateModel
+    rasterDirection <- rasterTemplateModel
+    
+    # Computing coordinates of raster
+    crds <- terra::crds(rasterWind, na.rm = FALSE)
+    
+    # Computing distances in degree to the eye of the storm for x and y axes
+    x <- crds[, 1] - dataTC$lon[j]
+    y <- crds[, 2] - dataTC$lat[j]
+    
+    # Computing distances to the eye of the storm in m
+    distEye <- terra::distance(
+      x = crds,
+      y = cbind(dataTC$lon[j], dataTC$lat[j]),
+      lonlat = TRUE
+    )
+    
+    # Computing wind speed/direction
+    output <- computeWindProfile(dataTC,
+                                 j,
+                                 method,
+                                 asymmetry,
+                                 x,
+                                 y,
+                                 crds,
+                                 distEye,
+                                 buffer,
+                                 sts@spatialLoiBuffer,
+                                 world,
+                                 indCountries)
+    
+    terra::values(rasterWind) <- output$wind
+    terra::values(rasterDirection) <- output$direction
+    
+    # Stacking products
+    if ("MSW" %in% product) {
[…]
