+++ R/spatialBehaviour.R
@@ -998,6 +998,9 @@ processingStorm <- function(storm,
                             tempRes,
                             spaceRes,
                             windThresholds,
+                            loi,
+                            world,
+                            indCountries,
                             rasterTemplate,
                             stackMSW,
                             stackPDI,
@@ -1066,7 +1069,7 @@ processingStorm <- function(storm,
                                  crds,
                                  distEye,
                                  buffer,
-                                 sts@spatialLoiBuffer,
+                                 loi,
                                  world,
                                  indCountries)
     
@@ -1452,6 +1455,9 @@ spatialBehaviour <- function(sts,
       tempRes = tempRes,
       spaceRes = spaceRes,
       windThresholds = windThreshold,
+      loi = sts@spatialLoiBuffer,
+      world = world,
+      indCountries = indCountries,
       rasterTemplate = rasterTemplate,
       stackMSW = finalStackMSW,
       stackPDI = finalStackPDI,
