### 2023-12-20 #5766 @hugoledoux
@editorial check references
