### 2023-12-20 #5766 @hugoledoux
@thomasarsouze all reviewers are happy and I also went over the paper and improves some sentences, see there: https://github.com/umr-amap/StormR/pull/78 (some are more stylistic so feel free to merge what you want).

After this, could you please:
