### 2023-12-20 #5766 @thomasarsouze
Thanks @hugoledoux

> @thomasarsouze all reviewers are happy and I also went over the paper and improves some sentences, see there: [umr-amap/StormR#78](https://github.com/umr-amap/StormR/pull/78) (some are more stylistic so feel free to merge what you want).

Suggestions merged.

>
> After this, could you please:
>
>     * [x]  Make a tagged release of your software, and list the version tag of the archived version here.
Tagged release named [JOSS-Release](https://github.com/umr-amap/StormR/releases/tag/JOSS-Release).
Other releases / tags are:
[v0.1.1](https://github.com/umr-amap/StormR/releases/tag/v0.1.1)
[v0.1.0](https://github.com/umr-amap/StormR/releases/tag/v0.1.0)

>
>     * [ x]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)

Archived in Zenodo

>
>     * [x]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
>
>     * [x]  Please list the DOI of the archived version here.
>
DOI JOSS-Release: https://zenodo.org/records/10410727
DOI v0.1.1: https://zenodo.org/records/10213689

>
> I can then move forward with recommending acceptance of the submission.
