### 2023-12-21 30ac706
Update R/temporalBehaviour.R
Co-authored-by: Thomas Arsouze <thomas.arsouze@cirad.fr>
