### 2023-12-21 cae5057
Update R/spatialBehaviour.R
Co-authored-by: Thomas Arsouze <thomas.arsouze@cirad.fr>
