+++ R/spatialBehaviour.R
@@ -1013,7 +1013,7 @@ maskProduct <- function(finalStack, loi, template) {
 #'   }
 #' @param windThreshold numeric vector. Minimal wind threshold(s) (in \eqn{m.s^{-1}}) used to
 #'   compute the duration of exposure when `product="Exposure"`. Default value is to set NULL, in this 
-#'   case, the windthresholds are the one used in the scale defined in the stormdatabase.
+#'   case, the windthresholds are the one used in the scale defined in the stromsList.
 #' @param method character. Model used to compute wind speed and direction.
 #' Three different models are implemented:
 #'   \itemize{
