### 2023-12-21 0fee561
Update tests/testthat/test-stormsList.R
Co-authored-by: Thomas Arsouze <thomas.arsouze@cirad.fr>
