### 2023-12-21 79b453a
Update vignettes/ExtractStorms.Rmd
Co-authored-by: Thomas Arsouze <thomas.arsouze@cirad.fr>
