### 2023-12-21 b912d80
Add example in vignette
