### 2023-12-21 8f06d52
Add example vignette
