+++ vignettes/ExtractStorms.Rmd
@@ -141,3 +141,16 @@ argument of the `plotStorms()` function as follows:
 ``` {r chunk12}
 plotStorms(st, category = c(4, 5), labels = FALSE, legends = "topright")
 ```
+
+
+#### Dynamic plot
+
+`plotStorms` allows the user to dynamically plot tracks within an interactive map using leaflet
+library by setting `dynamicPlot` to `TRUE`. Doing so, the user can explore the map the way he wants
+and click and each dotted colored observations to see there informations.
+
+``` {r chunk12}
+# Example of dynamic plot, using the same parameters above
+plotStorms(st, category = c(4, 5), labels = FALSE, legends = "topright", dynamicPlot=TRUE)
+```
+
