### 2023-12-21 #5766 @hugoledoux
@thomasarsouze I would recommend you sticking to major.minor.patch for the version, I think this is simpler for everyone (JOSS is then v0.1.2 or v0.2.0, up to you).

Also, the title on zenodo should match that of the paper exactly, this is important. And also the authors should be the same: right now the order is different and "Julien" is spelled with one "l" in zenodo and wiht two "l" on the paper.
