### 2023-12-21 #5766 @thomasarsouze
Ok, sorry for that.

I have done the modifications accordingly:
https://zenodo.org/records/10417763
and the tagged release is now v.0.1.2
