+++ R/spatialBehaviour.R
@@ -261,7 +261,7 @@ makeTemplateRaster <- function(buffer, res) {
 #' models
 #'
 #' @noRd
-#' @param rasterTemplate SpatRaster. Raster generated with makeTemplateRaster
+#' @param resolution  numeric. Raster resolution generated with makeTemplateRaster
 #'   function
 #' @param buffer numeric. Buffer size in degree
 #' @param data data.frame. Data generated with getInterpolatedData function
@@ -269,13 +269,13 @@ makeTemplateRaster <- function(buffer, res) {
 #'   generate raster
 #'
 #' @return a SpatRaster
-makeTemplateModel <- function(rasterTemplate, buffer, data, index) {
+makeTemplateModel <- function(resolution, buffer, data, index) {
   template <- terra::rast(
     xmin = data$lon[index] - buffer,
     xmax = data$lon[index] + buffer,
     ymin = data$lat[index] - buffer,
     ymax = data$lat[index] + buffer,
-    resolution = terra::res(rasterTemplate),
+    resolution = c(resolution,resolution),
     vals = NA,
     time = as.POSIXct(data$isoTimes[index])
   )
@@ -1230,10 +1230,8 @@ spatialBehaviour <- function(sts,
                              verbose = 2) {
   startTime <- Sys.time()
 
-  checkInputsSpatialBehaviour(
-    sts, product, windThreshold, method, asymmetry,
-    empiricalRMW, spaceRes, tempRes, verbose
-  )
+  checkInputsSpatialBehaviour(sts, product, windThreshold, method, asymmetry,
+                              empiricalRMW, spaceRes, tempRes, verbose)
 
 
   if (verbose > 0) {
@@ -1243,6 +1241,17 @@ spatialBehaviour <- function(sts,
 
   # Make raster template
   rasterTemplate <- makeTemplateRaster(sts@spatialLoiBuffer, resolutions[spaceRes])
+  
+  # templateExtent <- c(terra::ext(rasterTemplate)[1],
+  #                     terra::ext(rasterTemplate)[2],
+  #                     terra::ext(rasterTemplate)[3],
+  #                     terra::ext(rasterTemplate)[4])
+  # 
+  # templateOrigin <- terra::ext(rasterTemplate)
+  # templateRes <- resolutions[spaceRes]
+  
+  
+  
   # Buffer size in degree
   buffer <- 2.5
   # Initializing final raster stacks
@@ -1284,132 +1293,118 @@ spatialBehaviour <- function(sts,
     cat("  (", getNbStorms(sts), ") ", getNames(sts), "\n\n")
   }
 
-  for (st in sts@data) {
+  # Auto detect cores and prepare parallel processing
+  cl <- parallel::makeCluster(parallel::detectCores() - 1) #On laisse un coeur de dispo
+  doParallel::registerDoParallel(cl)
+  `%dopar%` <- foreach::`%dopar%`
+  
+  result <- foreach::foreach(
+    i = 1:length(sts@data),
+    .combine = 'c'
+  ) %dopar% {
+    
+    # Extract storm
+    storm = sts@data[[i]]
+    
+    # Buffer size in degree
+    buffer <- 2.5
+    
     # Handling indices inside loi.buffer or not
-    ind <- getIndices(st, 2, product)
+    ind <- getIndices(storm, 2, product)
 
-
-    it1 <- st@obs.all$iso.time[1]
-    it2 <- st@obs.all$iso.time[2]
+    it1 <- storm@obs.all$iso.time[1]
+    it2 <- storm@obs.all$iso.time[2]
     timeDiff <- as.numeric(as.POSIXct(it2) - as.POSIXct(it1))
[…]
