+++ R/spatialBehaviour.R
@@ -750,58 +750,6 @@ stackRaster <- function(stack, rasterTemplate, rasterWind) {
 
 
 
-
-
-#' Stack a computed PDI layer in a PDI raster stack
-#'
-#' @noRd
-#' @param stack list of SpatRaster. where to stack the layer
-#' @param rasterTemplate SpatRaster. Raster template generated with
-#'   makeTemplateRaster function
-#' @param rasterWind SpatRaster. Layer to add to the stack
-#'
-#' @return list of SpatRaster
-stackRasterPDI <- function(stack, rasterTemplate, rasterWind) {
-  rho <- 1
-  cd <- 0.002
-  # Raising to power 3
-  rasterWind <- rasterWind**3
-  # Applying both rho and surface drag coefficient
-  rasterWind <- rasterWind * rho * cd
-
-  return(stackRaster(stack, rasterTemplate, rasterWind))
-}
-
-
-
-
-
-#' Stack a computed Exposure layer in a Exposure raster stack
-#'
-#' @noRd
-#' @param stack list of SpatRaster. where to stack the layer
-#' @param rasterTemplate SpatRaster. Raster template generated with
-#'   makeTemplateRaster function
-#' @param rasterWind SpatRaster. Layer to add to the stack
-#' @param threshold numeric. Wind threshold
-#'
-#' @return list of SpatRaster
-stackRasterExposure <- function(stack, rasterTemplate, rasterWind, threshold) {
-  for (t in threshold) {
-    rasterCModel <- rasterWind
-    terra::values(rasterCModel) <- NA
-    ind <- which(terra::values(rasterWind) >= t)
-    rasterCModel[ind] <- 1
-    stack <- stackRaster(stack, rasterTemplate, rasterCModel)
-  }
-
-  return(stack)
-}
-
-
-
-
-
 #' Select the stack function to use depending on the product
 #'
 #' @noRd
@@ -814,12 +762,31 @@ stackRasterExposure <- function(stack, rasterTemplate, rasterWind, threshold) {
 #'
 #' @return list of SpatRaster
 stackProduct <- function(product, stack, rasterTemplate, rasterWind, threshold) {
+  
   if (product == "MSW") {
+    
     stack <- stackRaster(stack, rasterTemplate, rasterWind)
+    
   } else if (product == "PDI") {
+    
+    rho <- 1
+    cd <- 0.002
+    # Raising to power 3
+    rasterWind <- rasterWind**3
+    # Applying both rho and surface drag coefficient
+    rasterWind <- rasterWind * rho * cd
     stack <- stackRasterPDI(stack, rasterTemplate, rasterWind)
+    
   } else if (product == "Exposure") {
-    stack <- stackRasterExposure(stack, rasterTemplate, rasterWind, threshold)
+    
+    for (t in threshold) {
+      rasterCModel <- rasterWind
+      terra::values(rasterCModel) <- NA
+      ind <- which(terra::values(rasterWind) >= t)
+      rasterCModel[ind] <- 1
+      stack <- stackRaster(stack, rasterTemplate, rasterCModel)
+    }
+    
   }
 
   return(stack)
@@ -1230,10 +1197,8 @@ spatialBehaviour <- function(sts,
                              verbose = 2) {
   startTime <- Sys.time()
 
-  checkInputsSpatialBehaviour(
-    sts, product, windThreshold, method, asymmetry,
-    empiricalRMW, spaceRes, tempRes, verbose
-  )
+  checkInputsSpatialBehaviour(sts, product, windThreshold, method, asymmetry,
+                              empiricalRMW, spaceRes, tempRes, verbose)
 
 
   if (verbose > 0) {
[…]
