### 2023-12-27 f22cc8d
use terra wrap and unwrap for clustering computations
