### 2024-01-02 #5766 @kthyng
@thomasarsouze A JOSS review is usually a good time to bump to a nice round version number like v1.0.0 — but up to you.
