### 2024-01-08 0040c68
Fix last chunk + linting of ExtractStorms vignette
