### 2024-01-09 #5766 @thomasarsouze
@kthyng thanks for your inputs, they have been merged.

> A JOSS review is usually a good time to bump to a nice round version number like v1.0.0 — but up to you.

I prefer to stick to the current version, we have long-term developments that will later justify such a bump.
